% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layers.R
\name{bound_mech_low}
\alias{bound_mech_low}
\title{Mechanical internal boundary layer; lowest height.}
\usage{
bound_mech_low(dist)
}
\arguments{
\item{dist}{Distance to point of roughness change in m.}
}
\value{
Height of boundary layer in m.
}
\description{
Calculation of the lowest height of the mechanical internal boundary layer.
}
