% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layers.R
\name{bound_thermal_avg}
\alias{bound_thermal_avg}
\title{Thermal internal boundary layer.}
\usage{
bound_thermal_avg(ustar, v_a, temp_change_dist, pt_upwind, pt, lr)
}
\arguments{
\item{ustar}{Friction velocity u*.}

\item{v_a}{Windspeed in height of anemometer in m/s.}

\item{temp_change_dist}{Distance to point of temperature change in m.}

\item{pt_upwind}{Potential temperature in upwind direction in °C.}

\item{pt}{Potential temperature at site in °C.}

\item{lr}{Lapse rate in K/m (or °C/m)}
}
\value{

}
\description{
Calculation of the average height of the thermal internal boundary layer.
}
