% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{bowen_ratio}
\alias{bowen_ratio}
\title{Bowen ratio}
\usage{
bowen_ratio(t1_pot, dpot, dah)
}
\arguments{
\item{t1_pot}{potential temperature (see temp_pot_temp)}

\item{dpot}{difference in potential temperature between the two measurement
heights in degrees}

\item{dah}{difference in absolute humidity (kg/m^3) between the two measurement heights}
}
\value{
bowen ratio
}
\description{
Calculates bowen ratio
}
