% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_station.R
\name{build_weather_station}
\alias{build_weather_station}
\title{Weather Station}
\usage{
build_weather_station(
  lat = 50.8405027777778,
  lon = 8.68330333333333,
  elev = 270,
  surface_type = "Wiese",
  obs_height = 0.3,
  texture = "clay",
  valley = F,
  slope = NULL,
  z1,
  z2,
  depth1 = NULL,
  depth2 = NULL,
  datetime,
  t1,
  t2,
  v1,
  v2,
  hum1,
  hum2,
  p1 = NULL,
  p2 = NULL,
  rad_bal = NULL,
  sw_bal = NULL,
  lw_bal = NULL,
  albedo = NULL,
  soil_flux = NULL,
  ts1 = NULL,
  ts2 = NULL,
  moisture = NULL
)
}
\arguments{
\item{lat}{Latitude of location. Preset: 50.840502777777788.683303333333333 (climate station caldern).}

\item{lon}{Longitude of location. Preset: 8.683303333333333 (climate station caldern).}

\item{elev}{Elevation of location above sea level in m. Preset: 270 m (climate station caldern).}

\item{surface_type}{Surface Type. Form: Character string. One of: "Wiese", "Acker", "Gruenflaeche", "Strasse", "Landwirtschaft", "Siedlung", "Nadelwald", "Laubwald", "Mischwald", "Stadt". Preset: "Wiese.}

\item{obs_height}{Height of vegetation in m. Preset: 0.3.}

\item{texture}{Texture of ground. Form: Character string. One of: "clay", "sand". Preset: "clay".}

\item{valley}{TRUE, if climate station is positioned in a valley. Form: Boolean. Preset: FALSE.}

\item{slope}{Slope of hillside in \%. Form: Integer or Numeric. Preset: NULL.}

\item{z1}{Lower measurement height in m. Preset: 2m.}

\item{z2}{Upper measurement height in m. Preset: 2m.}

\item{depth1}{Upper depth of measurment in m. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{depth2}{Lower depth of measurment in m. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{datetime}{Name of datetime-coloumn in data. Form: Character string. NOTE: datetime needs to be converted POSIXlt-Format (see ?as.POSIXlt)}

\item{t1}{Vector that contains lower temperature data. Form: Character string.}

\item{t2}{Vector that contains upper temperature data. Form: Character string.}

\item{v1}{Vector that contains lower wind speed data. Form: Character string.}

\item{v2}{Vector that contains upper wind speed data. Form: Character string.}

\item{hum1}{Vector that contains lower humidity data. Form: Character string.}

\item{hum2}{Vector that contains upper humidity data. Form: Character string.}

\item{p1}{#Vector that contains lower pressure data. Form: Character string. Preset: NULL. Note: If NULL, pressure will be calculated.}

\item{p2}{#Vector that contains upper pressure data. Form: Character string. Preset: NULL. Note: If NULL, pressure will be calculated.}

\item{rad_bal}{Vector that contains total radiation balance. Form: Character string. Preset: NULL. Note: If NULL, rad_bal will be calculated (Albedo needed).}

\item{sw_bal}{Vector that contains shortwave radiation balance. Form: Character string. Preset: NULL. Note: If NULL, sw_bal will be calculated (Albedo needed).}

\item{lw_bal}{Vector that contains longwave radiation balance. Form: Character string. Preset: NULL. Note: If NULL, lw_bal will be calculated (Albedo needed).}

\item{albedo}{Vector that contains albedo. Form: Character string. Preset: NULL. Note: Only needed, if radiation balances shall be calculated.}

\item{soil_flux}{Vector that contains soil flux. Form: Character string. Preset: NULL. Note: If NULL, soil_flux will be calculated.}

\item{ts1}{Vector that contains upper ground temperature data. Form: Character string. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{ts2}{Vector that contains lower ground temperature data. Form: Character string. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{moisture}{Vector that ground moisture data. Form: Character string. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}
}
\value{
List containing:
\enumerate{
\item list of weather station location properties
\item list of weather station properties
\item list of weather station measurements, which will conatin NAs if they were not defined in the input
}
}
\description{
Creates a element, that contains all data regarding the weather station, its location and its measurements.
}
