% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_availability}
\alias{check_availability}
\title{Check Availability}
\usage{
check_availability(weather_station, ...)
}
\arguments{
\item{weather_station}{Object of class weather_station.}

\item{...}{Strings of properties to check.}
}
\value{
Absolutely nothing
}
\description{
Checks availability of passed properties in the weather station object.
If property is NULL, aborts with error.
}
