% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{gam}
\alias{gam}
\title{gamma PT coefficient}
\usage{
gam(t)
}
\arguments{
\item{t}{Air temperature in °C}
}
\value{
gamma coefficient for Priestley-Taylor calculations.
}
\description{
Calculates gamma for latent and sensible Priestley-Taylor-Method.
gamma is the temperature-sensitive psychrometer constant.
This function is a polynomial fit for the table 6 in Foken (2013), p.48.
}
