% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_absolute}
\alias{hum_absolute}
\title{Absolute humidity}
\usage{
hum_absolute(p_vapor, t_pot)
}
\arguments{
\item{p_vapor}{Vapor presure in hPa.}

\item{t}{Potential air temperature in Kelvin.}
}
\value{
Absolute humidity in kg/m^3.
}
\description{
Calculates absolute humidity from vapor pressure and temperature.
}
