% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_evap_heat}
\alias{hum_evap_heat}
\title{Heat of evaporation}
\usage{
hum_evap_heat(t)
}
\arguments{
\item{t}{Air temperature in degrees C.}
}
\value{
Heat of evaporation in J/kg.
}
\description{
Calculates heat of evaporation for water from air temperature.
}
