% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_moisture_gradient}
\alias{hum_moisture_gradient}
\title{Moisture gradient}
\usage{
hum_moisture_gradient(hum1, hum2, t1, t2, p1, p2, z1, z2)
}
\arguments{
\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{t1}{Air temperature at lower height in degrees C.}

\item{t2}{Air temperature at upper height in degrees C.}

\item{p1}{Air pressure at lower height in hPa.}

\item{p2}{Air pressure at lower height in hPa.}

\item{z1}{Lower measurement height in m.}

\item{z2}{Upper measurement height in m.}
}
\value{

}
\description{
Moisture gradient
}
