% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_precipitable_water}
\alias{hum_precipitable_water}
\title{Total precipitable water}
\usage{
hum_precipitable_water(p, t, elev)
}
\arguments{
\item{p}{Air pressure in hPa.}

\item{t}{Air temperature in degrees C.}

\item{elev}{Elevation above sea level in m.}
}
\value{
Total precipitable water in cm (grams).
}
\description{
Estimates total precipitable water in the atmosphere.
It uses a moist adiabatic temperature gradient which might not be
suitable for every application.
}
