% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_sat_vapor_pres}
\alias{hum_sat_vapor_pres}
\title{Saturation vapor pressure}
\usage{
hum_sat_vapor_pres(temp)
}
\arguments{
\item{temp}{numeric. Air temperature in degrees Celsius}
}
\value{
numeric. Saturation vapor pressure in hPa
}
\description{
Calculates the saturation vapor pressure from air temperature using the Magnus
formula (applicable over water surfaces).
}
