% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_specific}
\alias{hum_specific}
\title{Specific humidity}
\usage{
hum_specific(vapor_pressure, air_pressure)
}
\arguments{
\item{vapor_pressure}{numeric. Vapor pressure in hPa}

\item{air_pressure}{numeric. Air pressure in hPa}
}
\value{
numeric. Specific humidity in kg/kg
}
\description{
Calculates specfic humidity from vapor pressure and air pressure.
}
