% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_vapor_pres}
\alias{hum_vapor_pres}
\title{Vapor pressure}
\usage{
hum_vapor_pres(hum, t)
}
\arguments{
\item{hum}{Relative humidity in \%.}

\item{t}{Air temperature in degrees C.}
}
\value{
Vapor pressure in hPa.
}
\description{
Calculates vapor pressure from relative humidity and saturation vapor pressure
}
