% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{lat_penman}
\alias{lat_penman}
\title{Latent Heat Penman Method}
\usage{
lat_penman(datetime, v1, t1, h1, z1, rad_bal, elevation, lat, lon)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt)}

\item{v1}{Wind velocity (m/s)}

\item{t1}{Temperature (degrees celsius)}

\item{h1}{Relative humidity in percent (0-100)}

\item{z1}{Height of measurement for t1, v1}

\item{rad_bal}{Radiation balance (W/m^2)}

\item{elevation}{Elevation above sea level (m)}

\item{lat}{Latitude in decimal degrees}

\item{lon}{Longitude in decimal degrees}
}
\value{
Latent heat flux (W/m^2)
}
\description{
Calculates the latent heat flux using the Penman-Monteith equation. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
