% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{lat_priestley_taylor}
\alias{lat_priestley_taylor}
\title{Latent Heat Priestley-Taylor Method}
\usage{
lat_priestley_taylor(t1, rad_bal, soil_flux)
}
\arguments{
\item{t1}{Air temperature in degrees celsius}

\item{rad_bal}{Radiation balance (W/m^2)}

\item{soil_flux}{Soil flux (W/m^2)}
}
\value{
Latent heat flux (W/m^2)
}
\description{
Calculates the latent heat flux using the Priestley-Taylor method. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
