% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_monin}
\alias{latent_monin}
\title{Latent Heat using Monin-Obukhov length}
\usage{
latent_monin(air_density, monin, moist_gradient, ustar, ri, t1, z1)
}
\arguments{
\item{air_density}{Air density in kg*m^-3.}

\item{monin}{Monin-Obukhov length in m.}

\item{moist_gradient}{Specific moisture gradient (dq/dlog(z) with q in kg/kg)}

\item{ustar}{Friction velocity in m/s.}

\item{ri}{Gradient-Richardson-Number.}

\item{t1}{Air temperature in degrees C.}

\item{z1}{Height of air temperature measurment in m.}
}
\value{
Latent heat flux in W/m^2.
}
\description{
Calculates the latent heat flux using the Monin-Obukhov length. Negative
flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
