% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_penman}
\alias{latent_penman}
\title{Latent Heat Penman Method}
\usage{
latent_penman(datetime, v1, t1, h1, z1, rad_bal, elevation, lat, lon)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt).}

\item{v1}{Wind velocity in m/s.}

\item{t1}{Temperature in degrees C}

\item{h1}{Relative humidity in percent (0-100)}

\item{z1}{Height of measurement for t1, v1 in m.}

\item{rad_bal}{Radiation balance in W/m^2.}

\item{elevation}{Elevation above sea level in m.}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}
}
\value{
Latent heat flux in W/m^2.
}
\description{
Calculates the latent heat flux using the Penman-Monteith equation. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
