% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwave.R
\name{lw_atmospheric}
\alias{lw_atmospheric}
\title{Longwave radiation of the atmosphere}
\usage{
lw_atmospheric(emissivity_air, temp)
}
\arguments{
\item{emissivity_air}{numeric. Emissivity of the atmosphere (factor: 0-1)}

\item{temp}{numeric. Air temperature in degrees celsius}
}
\value{
numeric. Atmospheric radiation in W*m^-2
}
\description{
Longwave radiation of the atmosphere
}
