% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwave.R
\name{lw_emissivity_air}
\alias{lw_emissivity_air}
\title{Emissivity of the atmosphere}
\usage{
lw_emissivity_air(temp, elev, p)
}
\arguments{
\item{temp}{Air temperature in degrees C.}

\item{elev}{Elevation above sea level in m.}

\item{p}{Air pressure in hPa.}
}
\value{
Emissivity of the atmosphere (0-1).
}
\description{
Calculates emissivity of the atmosphere.
}
