% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwave.R
\name{lw_surface}
\alias{lw_surface}
\title{Longwave radiation of the surface}
\usage{
lw_surface(temp, emissivity = 0.95)
}
\arguments{
\item{temp}{Surface temperature in degrees C.}

\item{emissivity}{Emissivity of surface (0-1). Default is emissivity for short grass.}
}
\value{
Surface emissions in W/m^2.
}
\description{
Calculates emissions of a surface.
}
