% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{pos_anemometer_height}
\alias{pos_anemometer_height}
\title{Necessary Anemometer height.}
\usage{
pos_anemometer_height(dist, min_dist, obs_height)
}
\arguments{
\item{dist}{Distance between climate station and obstacle (e.g. forest) in m.}

\item{min_dist}{Minimum distance between climate station and obstacle, that is needed to ensure independency of measurements, in m.}

\item{obs_height}{Height of obstacle in m.}
}
\value{
Message, that tells you if climate station is well positioned or, if not, in which height the anemometer needs to be positioned to ensure independency of measurements.
}
\description{
Checks if distance of climate station to the forest is smaller than minimum distance. If so, calulates in which height the anemometer needs to be
positioned to ensure independency of measurements.
}
