% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position.R
\name{pos_max_dist}
\alias{pos_max_dist}
\title{Maximum distance between climate station and obstacle.}
\usage{
pos_max_dist(dist, obs_width, obs_height, ring = F)
}
\arguments{
\item{dist}{Distance between climate station and obstacle (e.g. forest) in m.}

\item{obs_width}{Width of obstacle in m.}

\item{obs_height}{Height of obstacle in m.}

\item{ring}{True, if obstacle is circularly surrounded by the obstacle.}
}
\value{
Message, that tells you if climate station is well positioned or, if not, in which distance to the obstacle it needs to be placed instead.
}
\description{
If station is positioned beyond minimum distance, check if it lies within maximum distance.
}
