% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priestley_taylor_coefficient}
\alias{priestley_taylor_coefficient}
\title{Priestley-Taylor coefficient for different surface types.}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{surface_type}{Type of surface as character}
\item{alpha}{coefficient value as numeric}
}
}
\usage{
priestley_taylor_coefficient
}
\description{
A dataset containing values for priestley-taylor coefficiencts for different surface types.
}
\keyword{datasets}
