% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_bal_total}
\alias{rad_bal_total}
\title{Total radiation balance}
\usage{
rad_bal_total(rad_sw_radiation_balance, rad_lw_surface, rad_lw_atmospheric)
}
\arguments{
\item{rad_sw_radiation_balance}{Shortwave radiation balance in W/m^2.}

\item{rad_lw_surface}{Longave surface emissions in W/m^2.}

\item{rad_lw_atmospheric}{Atmospheric radiation in W/m^2.}
}
\value{
Total radiation balance in W/m^2.
}
\description{
Calculate total radiation balance.
}
