% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_bal_total_with_topography}
\alias{rad_bal_total_with_topography}
\title{Total radiation balance with topography}
\usage{
rad_bal_total_with_topography(
  rad_sw_reflected_by_terrain,
  rad_lw_surface,
  rad_lw_atmospheric,
  terr_sky_view
)
}
\arguments{
\item{rad_sw_reflected_by_terrain}{Shortwave radiation balance in dependency of topography in W/m^2.}

\item{rad_lw_surface}{Longave surface emissions in W/m^2.}

\item{rad_lw_atmospheric}{Atmospheric radiation in W/m^2.}

\item{terr_sky_view}{Sky view factor from 0-1.}
}
\value{
Total radiation balance with topography in W/m^2.
}
\description{
Calculates the total radiation balance with topography.
}
