% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_in}
\alias{rad_diffuse_in}
\alias{rad_diffuse_in.default}
\alias{rad_diffuse_in.weather_station}
\title{Incoming diffused radiation}
\usage{
rad_diffuse_in(...)

\method{rad_diffuse_in}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  surface_type,
  slope,
  exposition,
  valley,
  ...
)

\method{rad_diffuse_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=trans_vapor]{trans_vapor}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=rad_sw_toa.default]{rad_sw_toa.default}}, \code{\link[=terr_sky_view]{terr_sky_view}}, \code{\link[=terr_terrain_angle.default]{terr_terrain_angle.default}}
  \describe{
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default 0.35.}
    \item{\code{sol_const}}{Solar constant in W/m\eqn{^2}, default 1368.}
  }}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}
}
\description{
Incoming diffused radiation
}
\references{
Bendix 2004, p. 58eq3.14, p55eq3.9
}
