% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_out}
\alias{rad_diffuse_out}
\alias{rad_diffuse_out.default}
\alias{rad_diffuse_out.weather_station}
\title{Diffused outgoing radiation}
\usage{
rad_diffuse_out(...)

\method{rad_diffuse_out}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  surface_type,
  slope,
  valley,
  ...
)

\method{rad_diffuse_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Provide \code{slope} and \code{exposition} to perform topographic correction.
}
\references{
p46eq3.3, p52eq3.8
}
