% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_in_topo}
\alias{rad_lw_in_topo}
\alias{rad_lw_in_topo.numeric}
\alias{rad_lw_in_topo.weather_station}
\title{Incoming longwave radiation corrected by topography.}
\usage{
rad_lw_in_topo(...)

\method{rad_lw_in_topo}{numeric}(rad_lw_out, rad_lw_in, terr_sky_view, ...)

\method{rad_lw_in_topo}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{rad_lw_out}{Longave surface emissions in W/m^2.}

\item{rad_lw_in}{Atmospheric radiation in W/m^2.}

\item{terr_sky_view}{Sky view factor from 0-1. (See \link{terr_sky_view})}

\item{weather_station}{Object of class weather_station.}
}
\value{
Incoming longwave radiation with topography in W/m^2.
}
\description{
Corrects the incoming longwave radiation using the sky view factor.
}
