% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_balance_topography}
\alias{rad_sw_balance_topography}
\alias{rad_sw_balance_topography.numeric}
\alias{rad_sw_balance_topography.weather_station}
\alias{rad_bal_total_with_topography.weather_station}
\title{Shortwave radiation balance in dependency of topography}
\usage{
rad_sw_balance_topography(...)

\method{rad_sw_balance_topography}{numeric}(
  slope,
  valley = F,
  sol_elevation,
  sol_azimuth,
  exposition = 0,
  rad_sw_ground_horizontal,
  albedo,
  trans_total
)

\method{rad_sw_balance_topography}{weather_station}(weather_station, exposition = 0, trans_total = 0.8)

\method{rad_sw_balance_topography}{weather_station}(weather_station, exposition = 0, trans_total = 0.8)
}
\arguments{
\item{slope}{Slope in degrees.}

\item{valley}{Is the climate station placed in a valley (TRUE/FALSE)?}

\item{sol_elevation}{Sun elevation in degrees.}

\item{sol_azimuth}{Sun azimuth in degrees.}

\item{exposition}{Exposition (North = 0, South = 180).}

\item{rad_sw_ground_horizontal}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{albedo}{Albedo of surface.}

\item{trans_total}{Total transmittance of the atmosphere (0-1).}

\item{weather_station}{Object of class weather_station.}
}
\value{
Shortwave radiation balance in dependency of topography in W/m^2.
}
\description{
Calculate shortwave radiation balance in dependency of topography.
}
