% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_sw_ground_horizontal}
\alias{rad_sw_ground_horizontal}
\title{Shortwave radiation onto a horizontal area on the ground.}
\usage{
rad_sw_ground_horizontal(rad_sw_toa, trans_total)
}
\arguments{
\item{rad_sw_toa}{Shortwave radiation at top of atmosphere in W/m^2.}

\item{trans_total}{Total transmittance (0-1)}
}
\value{
Shortwave radiation on the ground onto a horizontal area in W/m^2.
}
\description{
Calculation of the shortwave radiation onto a horizontal area on the ground.
}
