% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.numeric}
\alias{rad_sw_in.weather_station}
\title{Shortwave radiation onto a horizontal area on the ground}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{numeric}(
  datetime,
  lat,
  lon,
  elev,
  t,
  slope,
  exposition,
  sol_const = 1368,
  ozone_column = 0.35,
  vis = 30
)

\method{rad_sw_in}{weather_station}(weather_station, trans_total = NULL, oz = 0.35, vis = 30, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.}

\item{vis}{OPTIONAL. Needed if trans_total = NULL. Meteorological visibility in km.
Default is the visibility on a clear day.}

\item{weather_station}{Object of class weather_station.}

\item{trans_total}{Total transmittance of the atmosphere.}

\item{oz}{OPTIONAL. Needed if trans_total = NULL. Columnar ozone in cm.
Default is average global value.}

\item{rad_sw_toa}{Shortwave radiation at top of atmosphere in W/m².}
}
\value{
Shortwave radiation on the ground onto a horizontal area in W/m².

W/m²
}
\description{
Calculation of the shortwave radiation onto a horizontal area on the ground.
}
\references{
p46eq3.3
}
