% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.default}
\alias{rad_sw_in.weather_station}
\title{Shortwave incoming radiation}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{default}(datetime, lon, lat, elev, temp, ...)

\method{rad_sw_in}{weather_station}(weather_station, trans_total = NULL, oz = 0.35, vis = 30, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rad_sw_toa.default]{rad_sw_toa.default}}, \code{\link[=trans_gas.default]{trans_gas.default}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=trans_aerosol.default]{trans_aerosol.default}}, \code{\link[=terr_terrain_angle.default]{terr_terrain_angle.default}}
  \describe{
    \item{\code{sol_const}}{Solar constant in W/m\eqn{^2}}
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm. Default is average value of 0.35 cm.}
    \item{\code{slope}}{Slope in degree.}
    \item{\code{exposition}}{Exposition in degree.}
  }}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}

\item{weather_station}{Object of class weather_station.}

\item{trans_total}{Total transmittance of the atmosphere.}

\item{oz}{OPTIONAL. Needed if trans_total = NULL. Columnar ozone in cm.
Default is average global value.}

\item{vis}{OPTIONAL. Needed if trans_total = NULL. Meteorological visibility in km.
Default is the visibility on a clear day.}
}
\value{
W/m\eqn{^2}.
}
\description{
Provide \code{slope} and \code{exposition} to perform topographic correction.
}
\references{
p46eq3.3, p52eq3.8
}
