% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in_topo}
\alias{rad_sw_in_topo}
\alias{rad_sw_in_topo.numeric}
\alias{rad_sw_in_topo.weather_station}
\title{Incoming shortwave radiation in dependency of topography}
\usage{
rad_sw_in_topo(...)

\method{rad_sw_in_topo}{numeric}(
  slope,
  exposition = 0,
  terr_sky_view,
  sol_elevation,
  sol_azimuth,
  rad_sw_in,
  albedo,
  trans_total = 0.8,
  ...
)

\method{rad_sw_in_topo}{weather_station}(weather_station, trans_total = 0.8, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{slope}{Slope in degrees.}

\item{exposition}{Exposition (North = 0, South = 180).}

\item{terr_sky_view}{Sky view factor (0-1).}

\item{sol_elevation}{Sun elevation in degrees.}

\item{sol_azimuth}{Sun azimuth in degrees.}

\item{rad_sw_in}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{albedo}{Albedo of surface.}

\item{trans_total}{Total transmittance of the atmosphere (0-1).
Default is average atmospheric transmittance.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Shortwave radiation balance in dependency of topography in W/m^2.
}
\description{
Calculate shortwave radiation balance in dependency of topography.
}
