% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_radiation_balance}
\alias{rad_sw_radiation_balance}
\alias{rad_sw_radiation_balance.numeric}
\alias{rad_sw_radiation_balance.weather_station}
\title{Shortwave radiation balance}
\usage{
rad_sw_radiation_balance(...)

\method{rad_sw_radiation_balance}{numeric}(rad_sw_ground_horizontal, rad_sw_reflected)

\method{rad_sw_radiation_balance}{weather_station}(
  weather_station,
  trans_total = NULL,
  oz = 0.35,
  vis = 30
)
}
\arguments{
\item{rad_sw_ground_horizontal}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{rad_sw_reflected}{Reflected shortwave ratdiation in W/m^2.}

\item{weather_station}{Object of class weather_station.}

\item{trans_total}{Total transmittance of the atmosphere.}

\item{oz}{OPTIONAL. Needed if trans_total = NULL. Columnar ozone in cm.
Default is average global value.}

\item{vis}{OPTIONAL. Needed if trans_total = NULL. Meteorological visibility in km.
Default is the visibility on a clear day.}
}
\value{
Shortwave radiation balance in W/m^2.
}
\description{
Calculation of the shortwave radiation balance.
}
