% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_sw_toa}
\alias{rad_sw_toa}
\title{Shortwave radiation at top of atmosphere.}
\usage{
rad_sw_toa(datetime, lat, lon)
}
\arguments{
\item{datetime}{Date and time.}

\item{lat}{Latitude of the place of the climate station in decimal degrees.}

\item{lon}{Longitude of the place of the climate station in decimal degrees.}
}
\value{
Shortwave radiation at top of atmosphere in W/m^2.
}
\description{
Calculation of the shortwave radiation at the top of the atmosphere.
}
