% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{sc}
\alias{sc}
\title{sc PT coefficient (???)}
\usage{
sc(t1)
}
\arguments{
\item{t1}{numeric. Air temperature in degrees celsius}
}
\value{
numeric. sc
}
\description{
sc PT coefficient (???)
}
