% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sens_monin}
\alias{sens_monin}
\title{Sensible Heat using Monin-Obukhov length}
\usage{
sens_monin(air_density, monin, t_gradient, ustar, ri, z1)
}
\arguments{
\item{air_density}{Air density in (kg*m^-3)}

\item{monin}{Monin-Obukhov length (m)}

\item{t_gradient}{Specific moisture gradient (dq/dlog(z) with q in kg/kg)}

\item{ustar}{Friction velocity (m/s)}

\item{ri}{Gradient-Richardson-Number}

\item{z1}{Height of air temperature measurment (m)}
}
\value{
Sensible heat flux (W/m^2)
}
\description{
Calculates the sensible heat flux using the Monin-Obukhov length. Negative
flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
