% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sens_priestley_taylor}
\alias{sens_priestley_taylor}
\title{Sensible Heat Priestley-Taylor Method}
\usage{
sens_priestley_taylor(t1, rad_bal, soil_flux)
}
\arguments{
\item{t1}{Air temperature in degrees celsius}

\item{rad_bal}{Radiation balance (W/m^2)}

\item{soil_flux}{Soil flux (W/m^2)}
}
\value{
Sensible heat flux (W/m^2)
}
\description{
Calculates the Sensible heat flux using the Priestley-Taylor method. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
