% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sensible_bowen}
\alias{sensible_bowen}
\alias{sensible_bowen.numeric}
\alias{sensible_bowen.weather_station}
\title{Sensible Heat using Bowen Method}
\usage{
sensible_bowen(...)

\method{sensible_bowen}{numeric}(
  t1,
  t2,
  hum1,
  hum2,
  p1,
  p2,
  z1 = 2,
  z2 = 10,
  rad_bal,
  soil_flux,
  ...
)

\method{sensible_bowen}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in degrees C.}

\item{t2}{Temperature at upper height in degrees C.}

\item{hum1}{Relative humidity at lower height (e.g. height of anemometer) in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{p1}{Air pressure at lower height in hPa.}

\item{p2}{Air pressure at upper height in hPa.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{rad_bal}{Radiation balance in W/m^2.}

\item{soil_flux}{Soil flux in W/m^2.}

\item{weather_station}{Object of class weather_station}
}
\value{
Sensible heat flux in W/m^2
}
\description{
Calculates the sensible heat flux using the Bowen Method. Negative
flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
