% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_flux.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\title{Soil heat flux}
\usage{
soil_heat_flux(diff_temp, diff_depth, thermal_cond)
}
\arguments{
\item{diff_temp}{Difference in temperature between two depths.}

\item{diff_depth}{Difference in depths (m)}

\item{thermal_cond}{Thermal conductivity of soil (W/m K)}
}
\value{
Soil heat flux (W*m^-2)
}
\description{
Calculates soil heat flux from measurements in two different
depths and thermal conductivity of the soil.
}
\details{
Negative values
}
