% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_flux.R
\name{soil_thermal_cond}
\alias{soil_thermal_cond}
\title{Soil thermal conductivity}
\usage{
soil_thermal_cond(moisture, texture = "sand")
}
\arguments{
\item{moisture}{Numeric vector with soil moisture data (in Vol-\%)}

\item{texture}{Soil texture. Either "sand" or "clay".}
}
\value{
vector with soil thermal conductivity (W/m K)
}
\description{
Calculates soil thermal conductivity (W/m K) from soil moisture (Vol-\%) and texture.
}
\details{
Works by linearly interpolating thermal conductivity based on measured data.
}
