% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_angles}
\alias{sol_angles}
\title{Solar azimuth and elevation angles}
\usage{
sol_angles(datetime, lat, lon)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \code{\link[base:as.POSIX*]{as.POSIXlt}} and \link[base:strptime]{base::strptime} for conversion.}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}
}
\value{
data.frame with two columns: sol_azimuth and sol_elevation.
}
\description{
Calculates solar azimuth and solar elevation angle.
}
