% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_azimuth}
\alias{sol_azimuth}
\title{Solar azimuth angle}
\usage{
sol_azimuth(datetime, lat, lon)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt)}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}
}
\value{
Solar azimuth angle in degrees.
}
\description{
Solar azimuth angle
}
