% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_day_angle}
\alias{sol_day_angle}
\alias{sol_day_angle.numeric}
\title{Day angle}
\usage{
sol_day_angle(...)

\method{sol_day_angle}{numeric}(datetime)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{datetime in POSIXt}
}
\value{
degree
}
\description{
Day angle
}
