% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_declination}
\alias{sol_declination}
\alias{sol_declination.default}
\title{Solar declination}
\usage{
sol_declination(...)

\method{sol_declination}{default}(datetime, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}
}
\value{
degree
}
\description{
Solar declination
}
\references{
p243
}
