% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_ecliptic_length}
\alias{sol_ecliptic_length}
\alias{sol_ecliptic_length.default}
\title{Solar ecliptic length}
\usage{
sol_ecliptic_length(...)

\method{sol_ecliptic_length}{default}(datetime, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}
}
\value{
degree
}
\description{
Solar ecliptic length
}
\references{
p243
}
