% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_elevation}
\alias{sol_elevation}
\title{Solar elevation angle}
\usage{
sol_elevation(datetime, lat, lon)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt)}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}
}
\value{
Solar elevation angle in degrees.
}
\description{
Solar elevation angle
}
