% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_medium_anomaly}
\alias{sol_medium_anomaly}
\alias{sol_medium_anomaly.default}
\alias{sol_medium_anomaly.weather_station}
\title{Solar medium anomaly}
\usage{
sol_medium_anomaly(...)

\method{sol_medium_anomaly}{default}(datetime, ...)

\method{sol_medium_anomaly}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{weather_station}{Object of class weather_station.}
}
\value{
degree
}
\description{
Solar medium anomaly
}
\references{
p243
}
