% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_time_formula}
\alias{sol_time_formula}
\alias{sol_time_formula.default}
\alias{sol_time_formula.weather_station}
\title{Solar time formula}
\usage{
sol_time_formula(...)

\method{sol_time_formula}{default}(datetime, lon, ...)

\method{sol_time_formula}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{lon}{Longitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
hour
}
\description{
Solar time formula
}
\references{
p243
}
