% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surface_properties}
\alias{surface_properties}
\title{Emissivity and roughness length for different surfaces.}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{surface_type}{Type of surface as character}
\item{emissivity}{Emissivity of the chosen surface}
\item{roughness_length}{Roughness length of surface in m}
\item{albedo}{Albedo of the certain surface type as decimal number}
}
}
\usage{
surface_properties
}
\description{
A dataset containing values for emissivity and roughness length for different surface types.
}
\keyword{datasets}
