% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{temp_pot_temp}
\alias{temp_pot_temp}
\title{Potential Temperature.}
\usage{
temp_pot_temp(T1, p)
}
\arguments{
\item{T1}{Temperature in °C.}

\item{p}{Pressure in hPa.}
}
\value{

}
\description{
Calculation of the potential air temperature.
}
