% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terr_sky_view}
\alias{terr_sky_view}
\title{Sky view factor}
\usage{
terr_sky_view(slope, valley = F)
}
\arguments{
\item{slope}{Inclination of slope in degrees}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE)}
}
\value{
sky view factor from 0-1
}
\description{
Sky view factor
}
