% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terr_terrain_angle}
\alias{terr_terrain_angle}
\alias{terr_terrain_angle.default}
\alias{terr_terrain_angle.weather_station}
\title{Title}
\usage{
terr_terrain_angle(...)

\method{terr_terrain_angle}{default}(datetime, lon, lat, ..., slope = 0, exposition = 0)

\method{terr_terrain_angle}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{weather_station}{Object of class weather_station.}
}
\value{
degree.
}
\description{
Title
}
\references{
p52eq3.7
}
