% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_aerosol}
\alias{trans_aerosol}
\title{Transmittance due to aerosols}
\usage{
trans_aerosol(air_mass_abs, vis = 30)
}
\arguments{
\item{air_mass_abs}{Absolute optical air mass.}

\item{vis}{Meteorological visibility in km. Default is the visibility on a clear day.}
}
\value{
Transmittance due to aerosols (0-1).
}
\description{
Calculates transmittance due to aerosols.
}
