% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_abs}
\alias{trans_air_mass_abs}
\title{Absolute optical air mass}
\usage{
trans_air_mass_abs(air_mass_rel, pressure)
}
\arguments{
\item{air_mass_rel}{Relative optical air mass.}

\item{pressure}{Pressure in hPa.}
}
\value{
Absolute optical air mass.
}
\description{
Calculates absolute optical air mass.
}
