% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_gas}
\alias{trans_gas}
\title{Transmittance due to gas}
\usage{
trans_gas(air_mass_abs)
}
\arguments{
\item{air_mass_abs}{Absolute optical air mass.}
}
\value{
Transmittance due to gas (0-1)
}
\description{
Calculates transmittance due to O2 and CO2.
}
