% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_rayleigh}
\alias{trans_rayleigh}
\title{Transmittance due to rayleigh scattering}
\usage{
trans_rayleigh(air_mass_abs)
}
\arguments{
\item{air_mass_abs}{}
}
\value{

}
\description{
Transmittance due to rayleigh scattering
}
