% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_vapor}
\alias{trans_vapor}
\title{Transmittance due to water vapor}
\usage{
trans_vapor(air_mass_rel, pw)
}
\arguments{
\item{pw}{}
}
\value{

}
\description{
Transmittance due to water vapor
}
