% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence.R
\name{turb_displacement}
\alias{turb_displacement}
\title{Displacement height.}
\usage{
turb_displacement(obs_height)
}
\arguments{
\item{obs_height}{Height of obstacle in m.}
}
\value{
Displacement height in m.
}
\description{
Calculate the displacement height, caused by an obstacle (e.g. a crop field).
}
