% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_calc_fluxes}
\alias{turb_flux_calc_fluxes}
\title{Sensible and latent heat fluxes}
\usage{
turb_flux_calc_fluxes(weather_station)
}
\arguments{
\item{weather_station}{Object of class weather_station}
}
\value{
Object of class weather_station
}
\description{
Calculate sensible and latent heat fluxes, using the methods of Priestly-Taylor, Bowen, Monin and Penman (only latent).
}
