% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_ex_quotient_imp}
\alias{turb_flux_ex_quotient_imp}
\title{Exchange quotient for impulse transmission}
\usage{
turb_flux_ex_quotient_imp(grad_rich_no, ustar, monin, z1, air_density)
}
\arguments{
\item{grad_rich_no}{Gradient-Richardson-Number.}

\item{ustar}{Friction velocity in m/s.}

\item{monin}{Monin-Obhukov-Length in m.}

\item{z1}{Height in m.}

\item{air_density}{Air density in kg/m^3.}
}
\value{
Exchange quotient for impulse transmission in kg/(m*s).
}
\description{
Calculation of the exchange quotient of the turbulent impulse transmission.
}
