% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_grad_rich_no}
\alias{turb_flux_grad_rich_no}
\title{Gradient-Richardson-Number.}
\usage{
turb_flux_grad_rich_no(t1, t2, z1, z2, v1, v2, p1, p2)
}
\arguments{
\item{t1}{Temperature at lower height (e.g. height of anemometer) in °C.}

\item{t2}{Temperature at upper height in °C.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{p1}{Pressure at lower height (e.g. height of anemometer) in hPa.}

\item{p2}{Pressure at upper height in hPa.}
}
\value{
Gradient-Richardson-Number.
}
\description{
Calculation of the Gradient-Richardson-Number.
}
