% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_imp_exchange}
\alias{turb_flux_imp_exchange}
\title{Turbulent impulse exchange.}
\usage{
turb_flux_imp_exchange(ex_quotient, v1, v2, z1, z2)
}
\arguments{
\item{ex_quotient}{Exchange quotient in kg/(m*s).}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}
}
\value{
Turbulent impulse exchange in kg/(m*s^2).
}
\description{
Calculation of the turbulent impulse exchange.
}
