% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_monin}
\alias{turb_flux_monin}
\title{Monin-Obhukov-Length}
\usage{
turb_flux_monin(stability, z1 = 2, z2 = 10, z0, v1, v2, t1, t2)
}
\arguments{
\item{stability}{Stability of atmosphere.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{z0}{Roughness length in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in °C.}

\item{t2}{Temperature at upper height in °C.}

\item{ustar}{Friction velocity in m/s.}
}
\value{
Monin-Obhukov-Length in m.
}
\description{
Calculation of the Monin-Obhukov-Length.
Stability of atmosphere needs to be given as one of "stabil", "neutral" or "unstabil".
}
