% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_stability}
\alias{turb_flux_stability}
\title{Stability.}
\usage{
turb_flux_stability(grad_rich_no)
}
\arguments{
\item{grad_rich_no}{Gradient-Richardson-Number}
}
\value{
Stability of atmosphere. Returns one of "stabil", "neutral" or "instabil".
}
\description{
Calculation of atmospheric stability.
}
