% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence.R
\name{turb_roughness_length}
\alias{turb_roughness_length}
\title{Roughness length}
\usage{
turb_roughness_length(surface_type = NULL, obs_height = NULL)
}
\arguments{
\item{surface_type}{Type of surface}

\item{obs_height}{Height of obstacle in m.}
}
\value{
Roughness length in m.
}
\description{
Calculate the roughness length of an surface on the basis of the obstacle height or the type of the surface
Possible surface types are:
"Wiese", "Acker", "Gruenflaeche", "Strasse", "Landwirtschaft", "Siedlung", "Nadelwald", "Laubwald", "Mischwald", "Stadt"
You need to specify only one, "type" OR "obs_height".
}
