% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence.R
\name{turb_ustar}
\alias{turb_ustar}
\title{Friction velocity.}
\usage{
turb_ustar(v_a, z_a, z0)
}
\arguments{
\item{v_a}{Windspeed in height of anemometer in m/s.}

\item{z_a}{Height of anemometer in m.}

\item{z0}{Roughness length in m.}
}
\value{
Friction velocity in m/s.
}
\description{
Calculate the friction velocity of the underground.
}
