% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{as.data.frame.weather_station}
\alias{as.data.frame.weather_station}
\title{Create a data.frame from a weather station object}
\usage{
\method{as.data.frame}{weather_station}(x, ..., reduced = F, unit = F)
}
\arguments{
\item{x}{Object of class weather_station.}

\item{...}{Not used.}

\item{reduced}{TRUE, to only output the most important columns.}

\item{unit}{TRUE, to generate longer column labels with units.}
}
\value{
data.frame
}
\description{
Create a data.frame from a weather station object, that contains weather station measurements.
}
