% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layers.R
\name{bound_mech_avg}
\alias{bound_mech_avg}
\title{Mechanical internal boundary layer; average height.}
\usage{
bound_mech_avg(dist)
}
\arguments{
\item{dist}{Distance to point of roughness change in m.}
}
\value{
medium height of boundary layer in m.
}
\description{
Calculation of the average height of the mechanical internal boundary layer.
}
\references{
Bendix 2004, p. 242
}
