% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layers.R
\name{bound_thermal_avg}
\alias{bound_thermal_avg}
\title{Thermal internal boundary layer.}
\usage{
bound_thermal_avg(
  v,
  z,
  temp_change_dist,
  t_pot_upwind,
  t_pot,
  lapse_rate,
  surface_type = NULL,
  obs_height = NULL
)
}
\arguments{
\item{v}{Windspeed in height of anemometer in m/s.}

\item{z}{Height of anemometer in m.}

\item{temp_change_dist}{Distance to point of temperature change in m.}

\item{t_pot_upwind}{Potential temperature in upwind direction in °C.}

\item{t_pot}{Potential temperature at site in °C.}

\item{lapse_rate}{Lapse rate in °C/m.}

\item{surface_type}{Type of surface.}

\item{obs_height}{Height of obstacle in m.}
}
\value{
Average height of the thermal boundary layer in m.
}
\description{
Calculation of the average height of the thermal internal boundary layer.
}
\references{
Bendix 2004, p. 242
}
