% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_layers.R
\name{bound_thermal_avg}
\alias{bound_thermal_avg}
\title{Thermal internal boundary layer.}
\usage{
bound_thermal_avg(ustar, v, temp_change_dist, pt_upwind, t_pot, lapse_rate)
}
\arguments{
\item{ustar}{Friction velocity u* in m/s.}

\item{v}{Windspeed in height of anemometer in m/s.}

\item{temp_change_dist}{Distance to point of temperature change in m.}

\item{pt_upwind}{Potential temperature in upwind direction in °C.}

\item{t_pot}{Potential temperature at site in °C.}

\item{lr}{Lapse rate in K/m (or degrees C/m)}
}
\value{
Average height of the thermal boundary layer in m.
}
\description{
Calculation of the average height of the thermal internal boundary layer.
}
