% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{bowen_ratio}
\alias{bowen_ratio}
\title{Bowen-ratio}
\usage{
bowen_ratio(t, dpot, dah)
}
\arguments{
\item{t}{Air temperature in degrees C.}

\item{dpot}{Difference in potential temperature between the two measurement
heights in degrees Celsius.}

\item{dah}{Difference in absolute humidity (kg/m^3) between the two measurement heights.}
}
\value{
Bowen-ratio
}
\description{
Calculates Bowen-ratio.
}
