% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_station.R
\name{build_weather_station}
\alias{build_weather_station}
\title{Build a weather station object}
\usage{
build_weather_station(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rad_sw_toa.default]{rad_sw_toa.default}}, \code{\link[=pres_p.default]{pres_p.default}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=trans_aerosol.default]{trans_aerosol.default}}, \code{\link[=pres_sat_vapor_p.default]{pres_sat_vapor_p.default}}, \code{\link[=rad_lw_in.default]{rad_lw_in.default}}
  \describe{
    \item{\code{sol_const}}{Solar radiation constant in \eqn{W \cdot m^{-2}}. Default = 1368.}
    \item{\code{p0}}{Standard pressure in hPa, default 1013.25.}
    \item{\code{g}}{Gravitational acceleration in \eqn{m \cdot s^{-2}}, default 9.81.}
    \item{\code{rl}}{Specific gas constant for air in \eqn{m^2 \cdot s^{-2} \cdot K^{-1}}, default 287.05.}
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default 0.35.}
    \item{\code{vis}}{Visibility in km, default 30.}
    \item{\code{a}}{Constant a is 7.5 (default) over water, 7.6 over undercooled water, and 9.5 over ice.}
    \item{\code{b}}{Constant b is 235 (defalut) over water, 240.7 over undercooled water, and 265.5 over ice.}
    \item{\code{sigma}}{Stefan-Boltzmann constant in W/m\eqn{^2}/K\eqn{^4}, default `r sigma_default'.}
  }}

\item{weather_station}{Object of class \code{weather_station}.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{t1}{Air temperature at lower height in 째C.}

\item{t2}{Air temperature at upper height in 째C.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{obs_height}{Height of obstacle in m.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{surface_temp}{Surface temperature in \eqn{^\circ}C.}

\item{rh}{Relative humidity in \%.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{rad_bal}{Radiation balance in W/m\eqn{^2}.}

\item{texture}{Soil texture. Either \code{sand}, \code{peat} or \code{clay}.}

\item{moisture}{Soil moisture in cubic meter/cubic meter.}

\item{soil_flux}{Soil flux in W/m\eqn{^2}.}

\item{soil_temp1}{Soil temperature in 째C at upper height.}

\item{soil_temp2}{Soil temperature in 째C at lower height.}

\item{soil_depth}{Depth of the soil temperature measurement at upper height in m.}

\item{soil_depth2}{Depth of the soil temperature measurement at upper height in m.}
}
\description{
Creates a list of class \code{weather_station} that contains all input arguments.
}
\details{
Provided input arguments will only be used if they are listed in the section "Arguments".
}
