% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_station.R
\name{build_weather_station}
\alias{build_weather_station}
\title{Weather Station}
\usage{
build_weather_station(
  lat = 50.8405027777778,
  lon = 8.68330333333333,
  elev = 270,
  surface_type = "Wiese",
  obs_height = 0.3,
  texture = "clay",
  valley = F,
  slope = NULL,
  z1,
  z2,
  depth1 = NULL,
  depth2 = NULL,
  datetime,
  t1,
  t2,
  v1,
  v2,
  hum1,
  hum2,
  p1 = NULL,
  p2 = NULL,
  rad_bal = NULL,
  sw_bal = NULL,
  lw_bal = NULL,
  albedo = NULL,
  soil_flux = NULL,
  ts1 = NULL,
  ts2 = NULL,
  moisture = NULL
)
}
\arguments{
\item{lat}{Latitude of location. Preset: 50.84050277777778 (climate station caldern).}

\item{lon}{Longitude of location. Preset: 8.683303333333333 (climate station caldern).}

\item{elev}{Elevation of location above sea level in m. Preset: 270 m (climate station caldern).}

\item{surface_type}{Surface Type. Form: Character string. One of: "Wiese", "Acker", "Gruenflaeche", "Strasse", "Landwirtschaft", "Siedlung", "Nadelwald", "Laubwald", "Mischwald", "Stadt". Preset: "Wiese.}

\item{obs_height}{Height of vegetation in m. Preset: 0.3.}

\item{texture}{Texture of ground. Form: Character string. One of: "clay", "sand". Preset: "clay".}

\item{valley}{TRUE, if climate station is positioned in a valley. Form: logical. Preset: FALSE.}

\item{slope}{Slope of hillside in \%. Form: Numeric. Preset: NULL.}

\item{z1}{Lower measurement height in m. Preset: 2m.}

\item{z2}{Upper measurement height in m. Preset: 2m.}

\item{depth1}{Upper depth of soil measurment in m. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{depth2}{Lower depth of soil measurment in m. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{datetime}{Name of datetime-coloumn in data.
Form: POSIX-Object (See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.)}

\item{t1}{Vector containing lower temperature data in degrees C.}

\item{t2}{Vector containing upper temperature data in degrees C.}

\item{v1}{Vector containing lower wind speed data in m/s.}

\item{v2}{Vector containing upper wind speed data in m/s.}

\item{hum1}{Vector containing lower humidity data in \%.}

\item{hum2}{Vector containing upper humidity data in \%.}

\item{p1}{Vector containing lower pressure data in hPa.
Preset: NULL. Note: If NULL, pressure will be estimated.}

\item{p2}{Vector containing upper pressure data in hPa.
Preset: NULL. Note: If NULL, pressure will be calculated.}

\item{rad_bal}{Vector containing total radiation balance in W/m^2.
Preset: NULL. Note: If NULL, rad_bal will be calculated (Albedo needed).}

\item{sw_bal}{Vector containing shortwave radiation balance in W/m^2.
Preset: NULL. Note: If NULL, sw_bal will be calculated (Albedo needed).}

\item{lw_bal}{Vector containing longwave radiation balance in W/m^2.
Preset: NULL. Note: If NULL, lw_bal will be calculated (Albedo needed).}

\item{albedo}{Vector containing albedo.
Preset: NULL. Note: Only needed, if radiation balances shall be calculated.}

\item{soil_flux}{Vector containing soil flux in W/m^2.
Preset: NULL. Note: If NULL, soil_flux will be calculated.}

\item{ts1}{Vector containing upper ground temperature data in degrees C.
Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{ts2}{Vector containing lower ground temperature data in degrees C.
Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{moisture}{Vector containing ground moisture data in Vol-\%.
Preset: NULL. Note: Only needed, if soil flux shall be calculated.}
}
\value{
List of class "weather_station", that contains:
\enumerate{
\item list of location properties
\item list of weather station properties
\item list of measurements, which will conatin NAs if they were not defined in the input
}
}
\description{
Creates a list of class "weather_station, that contains all data regarding the weather station, its location and its measurements.
}
