% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_station.R
\name{build_weather_station}
\alias{build_weather_station}
\title{Weather Station}
\usage{
build_weather_station(
  lat = 50.840503,
  lon = 8.6833,
  elev = 270,
  surface_type = "Wiese",
  obs_height = 0.3,
  z1,
  z2,
  datetime,
  t1,
  t2,
  v1,
  v2,
  hum1,
  hum2,
  p1 = NULL,
  p2 = NULL,
  sw_in = NULL,
  sw_out = NULL,
  lw_in = NULL,
  lw_out = NULL,
  soil_flux = NULL,
  ...
)
}
\arguments{
\item{lat}{Latitude of location. Preset: 50.840503 (climate station caldern).}

\item{lon}{Longitude of location. Preset: 8.683300 (climate station caldern).}

\item{elev}{Elevation of location above sea level in m. Preset: 270 m (climate station caldern).}

\item{surface_type}{Surface Type. Form: Character string. One of: "Wiese", "Acker", "Gruenflaeche", "Strasse", "Landwirtschaft", "Siedlung", "Nadelwald", "Laubwald", "Mischwald", "Stadt". Preset: "Wiese.}

\item{obs_height}{Height of vegetation in m. Preset: 0.3.}

\item{z1}{Lower measurement height in m. Preset: 2m.}

\item{z2}{Upper measurement height in m. Preset: 2m.}

\item{datetime}{Name of datetime-coloumn in data.
Form: POSIX-Object (See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.)}

\item{t1}{Vector containing lower temperature data in degrees C.}

\item{t2}{Vector containing upper temperature data in degrees C.}

\item{v1}{Vector containing lower wind speed data in m/s.}

\item{v2}{Vector containing upper wind speed data in m/s.}

\item{hum1}{Vector containing lower humidity data in \%.}

\item{hum2}{Vector containing upper humidity data in \%.}

\item{p1}{Vector containing lower pressure data in hPa.
Preset: NULL. Note: If NULL, pressure will be estimated.}

\item{p2}{Vector containing upper pressure data in hPa.
Preset: NULL. Note: If NULL, pressure will be calculated.}

\item{soil_flux}{Vector containing soil flux in W/m^2.
Preset: NULL. Note: If NULL, soil_flux will be calculated.}

\item{depth1}{Upper depth of soil measurment in m. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{depth2}{Lower depth of soil measurment in m. Preset: NULL. Note: Only needed, if soil flux shall be calculated.}

\item{rad_bal}{Vector containing total radiation balance in W/m^2.
Preset: NULL. Note: If NULL, rad_bal will be calculated (Albedo needed).}

\item{sw_bal}{Vector containing shortwave radiation balance in W/m^2.
Preset: NULL. Note: If NULL, sw_bal will be calculated (Albedo needed).}

\item{lw_bal}{Vector containing longwave radiation balance in W/m^2.
Preset: NULL. Note: If NULL, lw_bal will be calculated (Albedo needed).}
}
\value{
List of class "weather_station", that contains:
\enumerate{
\item list of location properties
\item list of weather station properties
\item list of measurements, which will conatin NAs if they were not defined in the input
}
}
\description{
Creates a list of class "weather_station, that contains all data regarding the
weather station, its location and its measurements.
}
\details{
Parameters with preset NULL can be estimated using calculations. However some additional
variables need to be passed for the estimation of some parameters.
For usage examples see the examples below.

If p1 and p2 are NULL, they will get estimated using the elevation and air temperature.

If sw_in is NULL, it will get estimated using TOA radiation and average
atmospheric transmission (See \link{rad_sw_in}).
By also setting slope, sky_view and exposition, sw_in will be topographically corrected
(see \link{rad_sw_in_topo}).

If sw_out is NULL, albedo needs to be set (See \link{rad_sw_out}).

If lw_in is NULL, it will get estimated using the air temperature and pressure.
(See \link{rad_lw_in}).
By also setting sky_view, sw_in will be topographically corrected
(see \link{rad_sw_in_topo}).

If lw_out is NULL, t_surface needs to be set (See \link{rad_sw_out}).

If soil_flux is NULL, ts1, ts2, depth1, depth2, moisture and texture need to be set.
(See \link{soil_heat_flux} and \link{soil_thermal_cond}).
}
