% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{heat_capacity}
\alias{heat_capacity}
\title{Volumetric heat capacity}
\usage{
heat_capacity(t)
}
\arguments{
\item{t}{Air temperature in °C.}
}
\value{
Heat capacity density in J/(K*m³)
}
\description{
Calculates volumetric heat capacity
}
