% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{heat_capacity}
\alias{heat_capacity}
\title{Heat capacity density}
\usage{
heat_capacity(t_pot)
}
\arguments{
\item{t_pot}{Potential temperature in Kelvin.}
}
\value{
Heat capacity density in J/m^3 K
}
\description{
Calculates heat capacity density (?)
}
