% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_absolute}
\alias{hum_absolute}
\alias{hum_absolute.default}
\alias{hum_absolute.weather_station}
\title{Absolute humidity}
\usage{
hum_absolute(...)

\method{hum_absolute}{default}(rh, temp, ...)

\method{hum_absolute}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{rh}{Relative humidity in \%.}

\item{temp}{Temperature in °C.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Absolute humidity in kg/m\eqn{^3}.
}
\description{
Calculates absolute humidity from vapor pressure and air temperature.
}
\references{
Bendix 2004, p. 262.
}
