% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_absolute}
\alias{hum_absolute}
\alias{hum_absolute.default}
\alias{hum_absolute.weather_station}
\title{Absolute humidity}
\usage{
hum_absolute(...)

\method{hum_absolute}{default}(hum, t, ...)

\method{hum_absolute}{weather_station}(weather_station, height, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{hum}{Relative humidity in \%.}

\item{t}{Air temperature in °C.}

\item{height}{Height of measurement. "lower" or "upper".}
}
\value{
Absolute humidity in kg/m\eqn{^3}.
}
\description{
Calculates absolute humidity from vapor pressure and air temperature.
}
\references{
Bendix 2004, p. 262.
}
