% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_evap_heat}
\alias{hum_evap_heat}
\alias{hum_evap_heat.default}
\alias{hum_evap_heat.weather_station}
\title{Enthalpy of vaporization}
\usage{
hum_evap_heat(...)

\method{hum_evap_heat}{default}(temp, ...)

\method{hum_evap_heat}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in °C.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Enthalpy of vaporization in J/kg.
}
\description{
Calculates heat of evaporation for water from air temperature.
}
\references{
Bendix 2004, p. 261.
}
