% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_moisture_gradient}
\alias{hum_moisture_gradient}
\alias{hum_moisture_gradient.numeric}
\alias{hum_moisture_gradient.weather_station}
\title{Moisture gradient}
\usage{
hum_moisture_gradient(...)

\method{hum_moisture_gradient}{numeric}(hum1, hum2, t1, t2, p1, p2, z1 = 2, z2 = 10, ...)

\method{hum_moisture_gradient}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{t1}{Air temperature at lower height in degrees C.}

\item{t2}{Air temperature at upper height in degrees C.}

\item{p1}{Air pressure at lower height in hPa.}

\item{p2}{Air pressure at lower height in hPa.}

\item{z1}{Lower measurement height in m.}

\item{z2}{Upper measurement height in m.}

\item{weather_station}{Object of class weather_station}
}
\value{
Moisture gradient.
}
\description{
Calculates moisture gradient.
}
