% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_precipitable_water}
\alias{hum_precipitable_water}
\alias{hum_precipitable_water.default}
\alias{hum_precipitable_water.weather_station}
\title{Precipitable water}
\usage{
hum_precipitable_water(...)

\method{hum_precipitable_water}{default}(datetime, lat, elev, temp, ...)

\method{hum_precipitable_water}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_p.default]{pres_p.default}}
  \describe{
    \item{\code{g}}{Gravitational acceleration in m/s\eqn{^2}, default 9.81.}
    \item{\code{rl}}{Specific gas constant for air in m\eqn{^2}/s\eqn{^2}/K, default 287.05.}
  }}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Total precipitable water in cm\eqn{\cdot}grams.
}
\description{
Selects reference temperature and pressure based on location and season.
Then calculates precipitable water.
}
\details{
Latitude <= 30 degrees is defined as tropic; <= 60 is temperate; others is subarctic.
Summer is defined as April to September in the northern hemisphere.
}
\references{
Bendix 2004, p. 246. Column name "subarctic_summer" and "subarctic_winter" were switched.
}
