% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_precipitable_water}
\alias{hum_precipitable_water}
\alias{hum_precipitable_water.default}
\alias{hum_precipitable_water.weather_station}
\title{Total precipitable water}
\usage{
hum_precipitable_water(...)

\method{hum_precipitable_water}{default}(datetime, lat, elev, temp, ...)

\method{hum_precipitable_water}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_p.default]{pres_p.default}}
  \describe{
    \item{\code{g}}{Gravitational acceleration in m/s\eqn{^2}, default 9.81.}
    \item{\code{rl}}{Specific gas constant for air in m\eqn{^2}/s\eqn{^2}/K, default 287.05.}
  }}

\item{datetime}{Datetime in POSIXt}

\item{lat}{Latitude}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Total precipitable water in cm (grams).
}
\description{
Estimates total precipitable water in the atmosphere.
It uses a moist adiabatic temperature gradient which might not be
suitable for every application.
}
\details{
Latitude <= 30 degrees are defined as tropic; <= 60 are temperate; others are subarctic.
Summer is defined as April to September in the northern hemisphere.
}
\references{
p246. Column name "subarctic_summer" and "subarctic_winter" were switched.
}
