% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_precipitable_water}
\alias{hum_precipitable_water}
\alias{hum_precipitable_water.default}
\alias{hum_precipitable_water.weather_station}
\title{Total precipitable water}
\usage{
hum_precipitable_water(...)

\method{hum_precipitable_water}{default}(datetime, lat, elev, temp, p0 = p0_default, ...)

\method{hum_precipitable_water}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime}

\item{lat}{Latitude}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}

\item{p0}{Standard pressure}

\item{weather_station}{Object of class weather_station.}

\item{height}{Height of measurement. "lower" or "upper".}
}
\value{
Total precipitable water in cm (grams).
}
\description{
Estimates total precipitable water in the atmosphere.
It uses a moist adiabatic temperature gradient which might not be
suitable for every application.
}
\references{
p246
}
