% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_precipitable_water}
\alias{hum_precipitable_water}
\alias{hum_precipitable_water.numeric}
\alias{hum_precipitable_water.weather_station}
\title{Total precipitable water}
\usage{
hum_precipitable_water(...)

\method{hum_precipitable_water}{numeric}(p, t, elev, ...)

\method{hum_precipitable_water}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{p}{Air pressure in hPa.}

\item{t}{Air temperature in degrees C.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{Object of class weather_station.}

\item{height}{Height of measurement. "lower" or "upper".}
}
\value{
Total precipitable water in cm (grams).
}
\description{
Estimates total precipitable water in the atmosphere.
It uses a moist adiabatic temperature gradient which might not be
suitable for every application.
}
