% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_sat_vapor_pres}
\alias{hum_sat_vapor_pres}
\title{Saturation vapor pressure}
\usage{
hum_sat_vapor_pres(t)
}
\arguments{
\item{t}{Air temperature in degrees C.}
}
\value{
Saturation vapor pressure in hPa.
}
\description{
Calculates the saturation vapor pressure from air temperature using the Magnus
formula (applicable over water surfaces).
}
