% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_sat_vapor_pres}
\alias{hum_sat_vapor_pres}
\alias{hum_sat_vapor_pres.numeric}
\alias{hum_sat_vapor_pres.weather_station}
\title{Saturation vapor pressure}
\usage{
hum_sat_vapor_pres(...)

\method{hum_sat_vapor_pres}{numeric}(t, ...)

\method{hum_sat_vapor_pres}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t}{Air temperature in degrees C.}

\item{weather_station}{Object of class weather_station.}

\item{height}{Height of measurement. "lower" or "upper".}
}
\value{
Saturation vapor pressure in hPa.
}
\description{
Calculates the saturation vapor pressure from air temperature using the Magnus
formula (applicable over water surfaces).
}
