% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_specific}
\alias{hum_specific}
\title{Specific humidity}
\usage{
hum_specific(p_vapor, p)
}
\arguments{
\item{p_vapor}{Vapor pressure in hPa.}

\item{p}{Air pressure in hPa.}
}
\value{
Specific humidity in kg/kg.
}
\description{
Calculates specfic humidity from vapor pressure and air pressure.
}
