% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_vapor_pres}
\alias{hum_vapor_pres}
\title{Vapor pressure}
\usage{
hum_vapor_pres(hum_rel, sat_vapor_pres)
}
\arguments{
\item{hum_rel}{numeric. Relative humidity in percent (0-100)}

\item{sat_vapor_pres}{numeric. Saturation vapor pressure in hPa}
}
\value{
numeric. Vapor pressure in hPa
}
\description{
Calculates vapor pressure from relative humidity and saturation vapor pressure
}
