% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{hum_vapor_pres}
\alias{hum_vapor_pres}
\title{Vapor pressure}
\usage{
hum_vapor_pres(hum_rel, t1)
}
\arguments{
\item{hum_rel}{numeric. Relative humidity in percent (0-100).}

\item{t1}{numeric. Air temperature in degrees C.}
}
\value{
numeric. Vapor pressure in hPa.
}
\description{
Calculates vapor pressure from relative humidity and saturation vapor pressure
}
