% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_turbulent_flux.R
\name{lamb}
\alias{lamb}
\title{lambda PT coefficient (???)}
\usage{
lamb(t1)
}
\arguments{
\item{t1}{numeric. Air temperature in degree celsius}
}
\value{
numeric. lambda
}
\description{
lambda PT coefficient (???)
}
