% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_bowen}
\alias{latent_bowen}
\alias{latent_bowen.default}
\alias{latent_bowen.weather_station}
\title{Latent Heat using Bowen Method}
\usage{
latent_bowen(...)

\method{latent_bowen}{default}(
  t1,
  t2,
  hum1,
  hum2,
  z1 = 2,
  z2 = 10,
  elev,
  rad_bal,
  soil_flux,
  ...
)

\method{latent_bowen}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t1}{Temperature at lower height in °C.}

\item{t2}{Temperature at upper height in °C.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{z1}{Lower height of measuremen in m.}

\item{z2}{Upper height of measurement in m.}

\item{elev}{Elevation above sea level in m.}

\item{rad_bal}{Radiation balance in W/m².}

\item{soil_flux}{Soil flux in W/m².}

\item{weather_station}{Object of class weather_station.}
}
\value{
Latent heat flux in W/m².
}
\description{
Calculates the latent heat flux using the Bowen Method. Negative
flux signifies flux away from the surface, positive values signify flux
towards the surface.
Values above 600 W/m² and below -600 W/m² will be recognized
as measurement mistakes and smoothed respectively.
}
\references{
p221eq9.21.
}
