% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_monin}
\alias{latent_monin}
\alias{latent_monin.default}
\alias{latent_monin.weather_station}
\title{Latent Heat using Monin-Obukhov length}
\usage{
latent_monin(...)

\method{latent_monin}{default}(
  hum1,
  hum2,
  t1,
  t2,
  v1,
  v2,
  z1 = 2,
  z2 = 10,
  elev,
  surface_type = NULL,
  obs_height = NULL,
  ...
)

\method{latent_monin}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{hum1}{Relative humidity at lower height in \%.}

\item{hum2}{Relative humidity at upper height in \%.}

\item{t1}{Air temperature at lower height in °C.}

\item{t2}{Air temperature at upper height in °C.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{z1}{Lower height of measurement in m.}

\item{z2}{Upper height of measurement in m.}

\item{elev}{Elevation above sea level in m.}

\item{surface_type}{Type of surface. Options: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub}

\item{obs_height}{Height of obstacle in m.}

\item{weather_station}{Object of class weather_station}
}
\value{
Latent heat flux in W/m\eqn{^2}.
}
\description{
Calculates the latent heat flux using the Monin-Obukhov length. Negative
flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
\references{
Bendix 2004, p. 77, eq.4.6

Foken 2016, p. 61, Tab. 2.10
}
