% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_penman}
\alias{latent_penman}
\title{Latent Heat Penman Method}
\usage{
latent_penman(datetime, v, t, hum, z = 2, rad_bal, elev, lat, lon)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt).}

\item{v}{Wind velocity in m/s.}

\item{t}{Temperature in degrees C}

\item{hum}{Relative humidity in \%.}

\item{z}{Height of measurement for t, v in m.}

\item{rad_bal}{Radiation balance in W/m^2.}

\item{elev}{Elevation above sea level in m.}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}
}
\value{
Latent heat flux in W/m^2.
}
\description{
Calculates the latent heat flux using the Penman-Monteith equation. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
