% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_penman}
\alias{latent_penman}
\alias{latent_penman.POSIXt}
\alias{latent_penman.weather_station}
\title{Latent Heat Penman Method}
\usage{
latent_penman(...)

\method{latent_penman}{POSIXt}(datetime, v, t, hum, z = 2, rad_bal, elev, lat, lon, ...)

\method{latent_penman}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.}

\item{v}{Wind velocity in m/s.}

\item{t}{Temperature in degrees C}

\item{hum}{Relative humidity in \%.}

\item{z}{Height of measurement for t, v in m.}

\item{rad_bal}{Radiation balance in W/m^2.}

\item{elev}{Elevation above sea level in m.}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station}
}
\value{
Latent heat flux in W/m^2.
}
\description{
Calculates the latent heat flux using the Penman-Monteith equation. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
