% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_priestley_taylor}
\alias{latent_priestley_taylor}
\alias{latent_priestley_taylor.numeric}
\alias{latent_priestley_taylor.weather_station}
\title{Latent Heat Priestley-Taylor Method}
\usage{
latent_priestley_taylor(...)

\method{latent_priestley_taylor}{numeric}(t, z, rad_bal, soil_flux, surface_type = "field", ...)

\method{latent_priestley_taylor}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t}{Air temperature in °C.}

\item{z}{elevation of measurement in m.}

\item{rad_bal}{Radiation balance in W/m².}

\item{soil_flux}{Soil flux in W/m².}

\item{surface_type}{Surface type, for which a Priestley-Taylor coefficient will be selected. Default is for short grass.}

\item{weather_station}{Object of class weather_station}
}
\value{
Latent heat flux in W/m².
}
\description{
Calculates the latent heat flux using the Priestley-Taylor method. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
