% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent.R
\name{latent_priestley_taylor}
\alias{latent_priestley_taylor}
\alias{latent_priestley_taylor.numeric}
\alias{latent_priestley_taylor.weather_station}
\title{Latent Heat Priestley-Taylor Method}
\usage{
latent_priestley_taylor(...)

\method{latent_priestley_taylor}{numeric}(t, rad_bal, soil_flux)

\method{latent_priestley_taylor}{weather_station}(weather_station)
}
\arguments{
\item{t}{Air temperature in degrees Celsius.}

\item{rad_bal}{Radiation balance in W/m^2.}

\item{soil_flux}{Soil flux in W/m^2.}

\item{weather_station}{Object of class weather_station}
}
\value{
Latent heat flux in W/m^2.
}
\description{
Calculates the latent heat flux using the Priestley-Taylor method. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
