% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwave.R
\name{lw_atmospheric}
\alias{lw_atmospheric}
\title{Longwave radiation of the atmosphere}
\usage{
lw_atmospheric(temp, emissivity_air)
}
\arguments{
\item{temp}{Air temperature in degrees C.}

\item{emissivity_air}{Emissivity of the atmosphere (0-1).
See \code{\link{lw_emmissivity_air}} for calculation.}
}
\value{
Atmospheric radiation in W/m^2.
}
\description{
Longwave radiation of the atmosphere
}
