% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longwave.R
\name{lw_surface}
\alias{lw_surface}
\title{Longwave terrestric emission}
\usage{
lw_surface(temp, emissivity = 0.95)
}
\arguments{
\item{temp}{numeric. Surface temperature in degrees celsius}

\item{emissivity}{numeric. Emissivity of surface. Default is emissivity for short grass}
}
\value{
Emissions in W*m^-2
}
\description{
Calculates emissions of a surface.
}
