% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_air_density}
\alias{pres_air_density}
\alias{pres_air_density.weather_station}
\alias{pres_air_density.default}
\title{Air density}
\usage{
pres_air_density(...)

\method{pres_air_density}{weather_station}(weather_station, ...)

\method{pres_air_density}{default}(elev, temp, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{weather_station}{Object of class \code{weather_station}.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}
}
\value{
Air density in kg/m\eqn{^3}.
}
\description{
Calculation of the air density.
}
