% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_air_density}
\alias{pres_air_density}
\alias{pres_air_density.weather_station}
\alias{pres_air_density.default}
\title{Air density}
\usage{
pres_air_density(...)

\method{pres_air_density}{weather_station}(weather_station, height = "lower", ...)

\method{pres_air_density}{default}(p, t, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{height}{"lower" or "upper"}

\item{p}{Pressure in hPa.}

\item{t}{Temperature in °C.}
}
\value{
Air density in kg/m\eqn{^3}.
}
\description{
Calculation of the air density.
}
