% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_air_density}
\alias{pres_air_density}
\title{Air density}
\usage{
pres_air_density(p, t)
}
\arguments{
\item{p}{Pressure in hPa.}

\item{t}{Temperature in degrees C.}
}
\value{
Air density in kg/m^3.
}
\description{
Calculation of the air density.
}
