% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_p}
\alias{pres_p}
\alias{pres_p.default}
\alias{pres_p.weather_station}
\title{Air pressure}
\usage{
pres_p(...)

\method{pres_p}{default}(elev, temp, ..., p0 = p0_default, g = g_default, rl = rl_default)

\method{pres_p}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{p0}{Standard pressure in hPa, default 1013.25.}

\item{g}{Gravitational acceleration in \eqn{m \cdot s^{-2}}, default 9.81.}

\item{rl}{Specific gas constant for air in \eqn{m^2 \cdot s^{-2} \cdot K^{-1}}, default 287.05.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
hPa.
}
\description{
Calculate pressure based on barometric formula.
}
\details{
The formula assumes that the temperature does not change with altitude.
The results should be satisfying enough for an elevation lower than 5 km.
}
\references{
Lente & Ősz 2020 eq. 5.
}
