% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_p}
\alias{pres_p}
\alias{pres_p.default}
\alias{pres_p.weather_station}
\title{Air pressure}
\usage{
pres_p(...)

\method{pres_p}{default}(elev, temp, p0 = p0_default, g = g_default, rl = rl_default, ...)

\method{pres_p}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}

\item{p0}{Standard pressure in hPa.}

\item{g}{Gravitational acceleration in m/s${2}$.}

\item{rl}{Specific gas constant for air in j/kg/K.}

\item{weather_station}{Object of class weather_station.}

\item{height}{'lower' or 'upper'}
}
\value{
Pressure in hPa.
}
\description{
Calculation of pressure as a function of the elevation above sea level.
}
\references{
Lente & Ősz 2020 eq5.
}
