% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_p}
\alias{pres_p}
\alias{pres_p.default}
\alias{pres_p.weather_station}
\title{Air pressure}
\usage{
pres_p(...)

\method{pres_p}{default}(elev, temp, p0 = 1013, g = 9.81, rl = 287.05, ...)

\method{pres_p}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{Object of class weather_station.}

\item{height}{'lower' or 'upper'}

\item{t}{Temperature in °C.}
}
\value{
Pressure in hPa.
}
\description{
Calculation of pressure as a function of the elevation above sea level.
}
\references{
Lente & Ősz 2020 eq5.
}
