% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_p}
\alias{pres_p}
\title{Pressure.}
\usage{
pres_p(elev, t1)
}
\arguments{
\item{elev}{Elevation above sea level in m.}

\item{t1}{Temperature in °C.}
}
\value{
Pressure in hPa.
}
\description{
Calculation of pressure as a function of height.
}
