% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_p}
\alias{pres_p}
\title{Pressure.}
\usage{
pres_p(z, T1)
}
\arguments{
\item{z}{Height in m.}

\item{T1}{Temperature in °C.}
}
\value{
Pressure in hPa.
}
\description{
Calculation of pressure as a function of height.
}
