% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_sat_vapor_p}
\alias{pres_sat_vapor_p}
\alias{pres_sat_vapor_p.default}
\alias{pres_sat_vapor_p.weather_station}
\title{Saturated vapor pressure}
\usage{
pres_sat_vapor_p(...)

\method{pres_sat_vapor_p}{default}(temp, ..., a = 7.5, b = 235)

\method{pres_sat_vapor_p}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in degree Celcius.}

\item{a}{Constant a is 7.5 (default) over water, 7.6 over undercooled water, and 9.5 over ice.}

\item{b}{Constant b is 235 (defalut) over water, 240.7 over undercooled water, and 265.5 over ice.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
hPa.
}
\description{
Saturated vapor pressure
}
\references{
Bendix 2004, p. 261.
}
