% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_vapor_p}
\alias{pres_vapor_p}
\alias{pres_vapor_p.default}
\alias{pres_vapor_p.weather_station}
\title{Vapor pressure}
\usage{
pres_vapor_p(...)

\method{pres_vapor_p}{default}(temp, rh, ...)

\method{pres_vapor_p}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_sat_vapor_p.default]{pres_sat_vapor_p.default}}
  \describe{
    \item{\code{a}}{Constant a is 7.5 (default) over water, 7.6 over undercooled water, and 9.5 over ice.}
    \item{\code{b}}{Constant b is 235 (defalut) over water, 240.7 over undercooled water, and 265.5 over ice.}
  }}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{rh}{Relative humidity in \%.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
hPa.
}
\description{
Calculates vapor pressure from relative humidity and saturation vapor pressure.
}
