% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressure.R
\name{pres_vapor_p}
\alias{pres_vapor_p}
\alias{pres_vapor_p.default}
\alias{pres_vapor_p.weather_station}
\title{Vapor pressure}
\usage{
pres_vapor_p(...)

\method{pres_vapor_p}{default}(temp, rh, ...)

\method{pres_vapor_p}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_sat_vapor_p]{pres_sat_vapor_p}}
  \describe{
    \item{\code{}}{}
  }}

\item{temp}{Air temperature in °C.}

\item{rh}{Relative humidity in \%.}

\item{weather_station}{Object of class weather_station.}

\item{height}{Height of measurement. "lower" or "upper".}
}
\value{
Vapor pressure in hPa.
}
\description{
Calculates vapor pressure from relative humidity and saturation vapor pressure.
}
