% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_bal}
\alias{rad_bal}
\alias{rad_bal.default}
\alias{rad_bal.weather_station}
\title{Total radiation balance}
\usage{
rad_bal(...)

\method{rad_bal}{default}(datetime, lon, lat, elev, temp, rh, surface_temp, ...)

\method{rad_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Total radiation balance in W/m\eqn{^2}.
}
\description{
Calculate total radiation balance.
}
\references{
p45eq3.1
}
