% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_bal}
\alias{rad_bal}
\alias{rad_bal.numeric}
\alias{rad_bal.weather_station}
\title{Total radiation balance}
\usage{
rad_bal(...)

\method{rad_bal}{numeric}()

\method{rad_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{weather_station}{Object of class weather_station.}

\item{rad_sw_radiation_balance}{Shortwave radiation balance in W/m².}

\item{rad_lw_out}{Longwave surface emissions in W/m².}

\item{rad_lw_in}{Atmospheric radiation in W/m².}
}
\value{
Total radiation balance in W/m².
}
\description{
Calculate total radiation balance.
}
\references{
p45eq3.1
}
