% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_in}
\alias{rad_diffuse_in}
\alias{rad_diffuse_in.default}
\title{Incoming diffused radiation}
\usage{
rad_diffuse_in(...)

\method{rad_diffuse_in}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  ...,
  p0 = 1013,
  ozone_column = 0.35,
  sol_const = 1368,
  valley = FALSE,
  slope = 0,
  exposition = 0
)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{ozone_column}{Atmospheric ozone as column in cm. Default is average value of 0.35 cm.}

\item{sol_const}{Solar constant in W/m$^{2}$}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{slope}{Inclination of slope in degrees.}

\item{exposition}{Exposition in degree.}
}
\value{
Description.
}
\description{
Incoming diffused radiation
}
\references{
p58eq3.14, p55eq3.9
}
