% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_in}
\alias{rad_diffuse_in}
\alias{rad_diffuse_in.default}
\title{Incoming diffused radiation}
\usage{
rad_diffuse_in(...)

\method{rad_diffuse_in}{default}(datetime, lon, lat, elev, temp, surface_type = "field", ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=trans_vapor]{trans_vapor}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=rad_sw_toa]{rad_sw_toa}}, \code{\link[=terr_sky_view]{terr_sky_view}}, \code{\link[=terr_terrain_angle]{terr_terrain_angle}}
  \describe{
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm. Default is average value of 0.35 cm.}
  }}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}
}
\value{
Description.
}
\description{
Incoming diffused radiation
}
\references{
p58eq3.14, p55eq3.9
}
