% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_in}
\alias{rad_diffuse_in}
\alias{rad_diffuse_in.default}
\alias{rad_diffuse_in.weather_station}
\title{Incoming diffused radiation}
\usage{
rad_diffuse_in(...)

\method{rad_diffuse_in}{default}(datetime, lon, lat, elev, temp, slope, exposition, valley, ...)

\method{rad_diffuse_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rad_sw_toa.default]{rad_sw_toa.default}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}
  \describe{
    \item{\code{sol_const}}{Solar radiation constant in \eqn{W \cdot m^{-2}}, default = 1368.}
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default 0.35.}
  }}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Air temperature in degree Celcius.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Diffused shortwave incoming radiation.
}
\references{
Bendix 2004, p. 58 eq. 3.14, p. 55 eq. 3.9.
}
