% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_in}
\alias{rad_diffuse_in}
\alias{rad_diffuse_in.default}
\alias{rad_diffuse_in.weather_station}
\title{Incoming diffused radiation}
\usage{
rad_diffuse_in(...)

\method{rad_diffuse_in}{default}(datetime, lon, lat, elev, temp, ..., surface_type = "field")

\method{rad_diffuse_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=trans_vapor]{trans_vapor}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=rad_sw_toa]{rad_sw_toa}}, \code{\link[=terr_sky_view]{terr_sky_view}}, \code{\link[=terr_terrain_angle]{terr_terrain_angle}}
  \describe{
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default 0.35.}
  }}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}
}
\description{
Incoming diffused radiation
}
\references{
p58eq3.14, p55eq3.9
}
