% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_diffuse_out}
\alias{rad_diffuse_out}
\alias{rad_diffuse_out.default}
\alias{rad_diffuse_out.weather_station}
\title{Diffused outgoing radiation}
\usage{
rad_diffuse_out(...)

\method{rad_diffuse_out}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  slope,
  exposition,
  valley,
  surface_type,
  ...
)

\method{rad_diffuse_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Reflected diffused incoming radiation.
}
\references{
Bendix 2004, p. 45 eq. 3.1.
}
