% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\alias{rad_emissivity_air.default}
\alias{rad_emissivity_air.weather_station}
\title{Emissivity of the atmosphere}
\usage{
rad_emissivity_air(...)

\method{rad_emissivity_air}{default}(temp, rh, ...)

\method{rad_emissivity_air}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_vapor_p.default]{pres_vapor_p.default}}
  \describe{
    \item{\code{}}{}
  }}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{rh}{Relative humidity in \%.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Uniteless from 0 to 1.
}
\description{
Calculation of the emissivity of the atmosphere.
}
\references{
Bendix 2004, p. 66 eq. 3.22.
}
