% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\alias{rad_emissivity_air.default}
\alias{rad_emissivity_air.weather_station}
\title{Emissivity of the atmosphere}
\usage{
rad_emissivity_air(...)

\method{rad_emissivity_air}{default}(temp, rh, ...)

\method{rad_emissivity_air}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_sat_vapor_p.default]{pres_sat_vapor_p.default}}
  \describe{
    \item{\code{a}}{Constant a is 7.5 (default) over water, 7.6 over undercooled water, and 9.5 over ice.}
    \item{\code{b}}{Constant b is 235 (defalut) over water, 240.7 over undercooled water, and 265.5 over ice.}
  }}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
How much does the air emit longwave radiation compared to a black body?
}
\references{
Bendix 2004, p. 66 eq. 3.22.
}
