% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\alias{rad_emissivity_air.default}
\alias{rad_emissivity_air.weather_station}
\title{Emissivity of the atmosphere}
\usage{
rad_emissivity_air(...)

\method{rad_emissivity_air}{default}(temp, rh, ...)

\method{rad_emissivity_air}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_p]{pres_p}}, \code{\link[=pres_sat_vapor_p]{pres_sat_vapor_p}}
  \describe{
    \item{\code{}}{}
  }}

\item{temp}{Temperature in degree Celcius.}

\item{rh}{relative humidity.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Emissivity of the atmosphere (0-1).
}
\description{
Calculation of the emissivity of the atmosphere.
}
\references{
p67eq3.22
}
