% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\alias{rad_emissivity_air.default}
\alias{rad_emissivity_air.weather_station}
\title{Emissivity of the atmosphere}
\usage{
rad_emissivity_air(...)

\method{rad_emissivity_air}{default}(temp, rh, elev = 0, p = p0, p0 = 1013, ...)

\method{rad_emissivity_air}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in °C.}

\item{rh}{relative humidity.}

\item{elev}{Meters above sea level in m.}

\item{p}{OPTIONAL. Air pressure in hPa.}

\item{p0}{Standard air pressure in hPa.}

\item{weather_station}{Object of class weather_station.}

\item{height}{Height of measurement. "lower" or "upper".}
}
\value{
Emissivity of the atmosphere (0-1).
}
\description{
Calculation of the emissivity of the atmosphere.
}
\references{
p67eq3.23, 3.24, 3.25
}
