% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\title{Emissivity of the atmosphere}
\usage{
rad_emissivity_air(t, elev, p = NULL)
}
\arguments{
\item{t}{Air temperature in degrees C.}

\item{elev}{Meters above sea level}

\item{p}{Optional. Air pressure in hPa.
If not available, will be calculated from elev and air temperature.}
}
\value{
missivity of the atmosphere (0-1).
}
\description{
Calculation of the emissivity of the atmosphere.
}
