% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_emissivity_air}
\alias{rad_emissivity_air}
\title{Emissivity of the atmosphere}
\usage{
rad_emissivity_air(temp, altitude, air_pressure = NULL)
}
\arguments{
\item{temp}{numeric. Air termperature in degrees celsius.}

\item{altitude}{numeric. Meters above sea level}

\item{air_pressure}{numeric. Optional. Air pressure in hPa.
If not available, will be calculated from altitude and air temperature.}
}
\value{
numeric. Emissivity of the atmosphere (0-1)
}
\description{
Emissivity of the atmosphere
}
