% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_lw_atmospheric}
\alias{rad_lw_atmospheric}
\title{Longwave radiation of the atmosphere}
\usage{
rad_lw_atmospheric(emissivity_air, t)
}
\arguments{
\item{emissivity_air}{Emissivity of the atmosphere (factor: 0-1)}

\item{t}{Air temperature in degrees C.}
}
\value{
Atmospheric radiation in W/m^2.
}
\description{
Calculation of the longwave radiation of the atmosphere.
}
