% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_bal}
\alias{rad_lw_bal}
\alias{rad_lw_bal.default}
\title{Long wave radiation balance}
\usage{
rad_lw_bal(...)

\method{rad_lw_bal}{default}(name, ...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
W/m$^{2}$.
}
\description{
Long wave radiation balance
}
\references{
p68eq3.25
}
