% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_bal}
\alias{rad_lw_bal}
\alias{rad_lw_bal.default}
\alias{rad_lw_bal.weather_station}
\title{Long wave radiation balance}
\usage{
rad_lw_bal(...)

\method{rad_lw_bal}{default}(temp, rh, slope, valley, surface_type, surface_temp, ...)

\method{rad_lw_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rad_lw_in.default]{rad_lw_in.default}}
  \describe{
    \item{\code{sigma}}{Stefan-Boltzmann constant in \eqn{W \cdot m{^-2} \cdot K{^-4}},
default 5.6704 × 10\if{html}{\out{<sup>}}-8\if{html}{\out{</sup>}}.}
  }}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{slope}{Slope in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{surface_type}{Surface type.
Allowed values are: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub.
Object \code{surface_properties} is a table of surface properties.}

\item{surface_temp}{Surface temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Sum of longwave incoming and outgoing radiation.
}
\references{
Bendix 2004, p. 68. eq. 3.25.
}
