% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_bal}
\alias{rad_lw_bal}
\alias{rad_lw_bal.default}
\alias{rad_lw_bal.weather_station}
\title{Long wave radiation balance}
\usage{
rad_lw_bal(...)

\method{rad_lw_bal}{default}(temp, rh, surface_temp, ...)

\method{rad_lw_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rad_lw_in.default]{rad_lw_in.default}}, \code{\link[=rad_lw_out.default]{rad_lw_out.default}}
  \describe{
    \item{\code{sigma}}{Stefan-Boltzmann constant in W/m\eqn{^2}/K\eqn{^4},
default `r sigma_default'.}
    \item{\code{surface_type}}{Surface type for which a specific emissivity will be selected.}
  }}

\item{temp}{Temperature in degree Celcius.}

\item{rh}{relative humidity.}

\item{surface_temp}{Surface temperature in °C.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Long wave radiation balance
}
\references{
p68eq3.25
}
