% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_in}
\alias{rad_lw_in}
\alias{rad_lw_in.default}
\alias{rad_lw_in.weather_station}
\title{Longwave incoming radiation}
\usage{
rad_lw_in(...)

\method{rad_lw_in}{default}(temp, rh, slope, valley, ..., sigma = sigma_default)

\method{rad_lw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in degree Celcius.}

\item{rh}{Relative humidity in \%.}

\item{slope}{Slope in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{sigma}{Stefan-Boltzmann constant in \eqn{W \cdot m{^-2} \cdot K{^-4}},
default 5.6704 × 10\if{html}{\out{<sup>}}-8\if{html}{\out{</sup>}}.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Longwave radiation of the atmosphere.
}
\references{
Bendix 2004, p. 68 eq. 3.24. The second part of the equation
related to the surrounding terrain is not included.
}
