% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_in}
\alias{rad_lw_in}
\alias{rad_lw_in.default}
\alias{rad_lw_in.weather_station}
\title{Longwave radiation of the atmosphere}
\usage{
rad_lw_in(...)

\method{rad_lw_in}{default}(temp, rh, sigma = sigma_default, ...)

\method{rad_lw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=terr_sky_view.default]{terr_sky_view.default}}
  \describe{
    \item{\code{slope}}{Inclination of slope in degrees.}
    \item{\code{valley}}{If the position is in a valley (TRUE) or on a slope (FALSE).}
  }}

\item{temp}{Temperature in °C.}

\item{rh}{relative humidity.}

\item{sigma}{Stefan-Boltzmann constant in W/m\eqn{^2}/K\eqn{^4},
default `r sigma_default'.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Atmospheric radiation in W/m\eqn{^2}.
}
\description{
Calculation of the longwave radiation of the atmosphere.
}
\details{
The second part of the equation related to the surrounding terrain is not included.
}
\references{
p66eq3.24
}
