% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_in}
\alias{rad_lw_in}
\alias{rad_lw_in.numeric}
\alias{rad_lw_in.weather_station}
\title{Longwave radiation of the atmosphere}
\usage{
rad_lw_in(...)

\method{rad_lw_in}{numeric}(emissivity_air, t, ...)

\method{rad_lw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{emissivity_air}{Emissivity of the atmosphere (factor: 0-1)}

\item{t}{Air temperature in degrees C.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Atmospheric radiation in W/m^2.
}
\description{
Calculation of the longwave radiation of the atmosphere.
}
