% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_out}
\alias{rad_lw_out}
\alias{rad_lw_out.default}
\alias{rad_lw_out.weather_station}
\title{Longwave radiation of the surface}
\usage{
rad_lw_out(...)

\method{rad_lw_out}{default}(surface_temp, ..., surface_type = "field", sigma = sigma_default)

\method{rad_lw_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{surface_temp}{Surface temperature in °C.}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{sigma}{Stefan-Boltzmann constant
with the default 5.6993e-8 W/m\eqn{^2}/K$^{4}$
Default is 'field' as surface type.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Emissions in W/m\eqn{^2}.
}
\description{
Calculates emissions of a surface.
}
\details{
If a weather_station object is given, the lower air temperature will be used
instead of the surface temperature.
}
\references{
p66eq3.20
}
