% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_lw_out}
\alias{rad_lw_out}
\alias{rad_lw_out.numeric}
\alias{rad_lw_out.weather_station}
\title{Longwave radiation of the surface}
\usage{
rad_lw_out(...)

\method{rad_lw_out}{numeric}(t_surface, emissivity_surface = 0.95, ...)

\method{rad_lw_out}{weather_station}(weather_station, emissivity_surface = 0.95, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t_surface}{Surface temperature in degrees C.}

\item{emissivity_surface}{Emissivity of surface.
Default is emissivity for short grass.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Emissions in W/m^2.
}
\description{
Calculates emissions of a surface.
}
\details{
If a weather_station object is given, the lower air temperature will be used
instead of the surface temperature.
}
