% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_lw_surface}
\alias{rad_lw_surface}
\title{Longwave radiation of the surface}
\usage{
rad_lw_surface(t, emissivity_surface = 0.95)
}
\arguments{
\item{t}{Surface temperature in degrees C.}

\item{emissivity_surface}{Emissivity of surface. Default is emissivity for short grass.}
}
\value{
Emissions in W/m^2.
}
\description{
Calculates emissions of a surface.
}
