% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_bal}
\alias{rad_sw_bal}
\alias{rad_sw_bal.default}
\alias{rad_sw_bal.weather_station}
\title{Shortwave radiation balance}
\usage{
rad_sw_bal(...)

\method{rad_sw_bal}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  surface_type,
  slope,
  exposition,
  valley,
  ...
)

\method{rad_sw_bal}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}
}
\description{
Shortwave radiation balance
}
\references{
Bendix 2004, p. 45eq3.1, p63eq3.18
}
