% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_bal}
\alias{rad_sw_bal}
\alias{rad_sw_bal.numeric}
\title{Shortwave radiation balance}
\usage{
rad_sw_bal(...)

\method{rad_sw_bal}{numeric}(...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}
}
\value{
W/m$^{2}$
}
\description{
Shortwave radiation balance
}
\references{
p45eq3.1, p63eq3.18
}
