% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.default}
\alias{rad_sw_in.weather_station}
\title{Shortwave incoming radiation}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  ...,
  sol_const = 1368,
  p0 = 1013,
  ozone_column = 0.35,
  vis = 30,
  slope = 0,
  exposition = 0
)

\method{rad_sw_in}{weather_station}(weather_station, trans_total = NULL, oz = 0.35, vis = 30, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{sol_const}{Solar constant in W/m$^{2}$}

\item{ozone_column}{Atmospheric ozone as column in cm. Default is average value of 0.35 cm.}

\item{vis}{OPTIONAL. Needed if trans_total = NULL. Meteorological visibility in km.
Default is the visibility on a clear day.}

\item{weather_station}{Object of class weather_station.}

\item{trans_total}{Total transmittance of the atmosphere.}

\item{oz}{OPTIONAL. Needed if trans_total = NULL. Columnar ozone in cm.
Default is average global value.}
}
\value{
W/m$^{2}$.
}
\description{
Shortwave incoming radiation
}
\references{
p46eq3.3, p52eq3.8
}
