% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.default}
\alias{rad_sw_in.weather_station}
\title{Shortwave incoming radiation}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  surface_type,
  slope,
  exposition,
  valley,
  ...,
  sol_const = sol_const_default
)

\method{rad_sw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=trans_aerosol.default]{trans_aerosol.default}}
  \describe{
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default 0.35.}
    \item{\code{vis}}{Visibility in km, default 30.}
  }}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{sol_const}{Solar constant in W/m\eqn{^2}, default 1368.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Provide \code{slope} and \code{exposition} to perform topographic correction.
}
\references{
Bendix 2004, p. 46 eq. 3.3, p. 52 eq. 3.8.
}
