% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.default}
\alias{rad_sw_in.weather_station}
\title{Shortwave incoming radiation}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{default}(datetime, lon, lat, elev, temp, ..., surface_type = "field")

\method{rad_sw_in}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=rad_sw_toa.default]{rad_sw_toa.default}}, \code{\link[=trans_ozone.default]{trans_ozone.default}}, \code{\link[=trans_aerosol.default]{trans_aerosol.default}}, \code{\link[=terr_terrain_angle.default]{terr_terrain_angle.default}}
  \describe{
    \item{\code{sol_const}}{Solar constant in W/m\eqn{^2}, default 1368.}
    \item{\code{ozone_column}}{Atmospheric ozone as column in cm, default 0.35.}
    \item{\code{vis}}{Visibility in km, default 30.}
    \item{\code{slope}}{Slope in degree.}
    \item{\code{exposition}}{Exposition in degree.}
  }}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Provide \code{slope} and \code{exposition} to perform topographic correction.
}
\references{
p46eq3.3, p52eq3.8
}
