% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_in}
\alias{rad_sw_in}
\alias{rad_sw_in.numeric}
\alias{rad_sw_in.weather_station}
\title{Shortwave radiation onto a horizontal area on the ground}
\usage{
rad_sw_in(...)

\method{rad_sw_in}{numeric}(rad_sw_toa, trans_total, ...)

\method{rad_sw_in}{weather_station}(weather_station, trans_total = NULL, oz = 0.35, vis = 30, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{rad_sw_toa}{Shortwave radiation at top of atmosphere in W/m^2.}

\item{trans_total}{Total transmittance of the atmosphere.}

\item{weather_station}{Object of class weather_station.}

\item{oz}{OPTIONAL. Needed if trans_total = NULL. Columnar ozone in cm.
Default is average global value.}

\item{vis}{OPTIONAL. Needed if trans_total = NULL. Meteorological visibility in km.
Default is the visibility on a clear day.}
}
\value{
Shortwave radiation on the ground onto a horizontal area in W/m^2.
}
\description{
Calculation of the shortwave radiation onto a horizontal area on the ground.
}
