% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_out}
\alias{rad_sw_out}
\alias{rad_sw_out.default}
\alias{rad_sw_out.weather_station}
\title{Shortwave outgoing radiation}
\usage{
rad_sw_out(...)

\method{rad_sw_out}{default}(
  datetime,
  lon,
  lat,
  elev,
  temp,
  slope,
  exposition,
  valley,
  surface_type,
  ...
)

\method{rad_sw_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{surface_type}{Surface type for which a specific emissivity will be selected.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Provide \code{slope} and \code{exposition} to perform topographic correction.
}
\references{
Bendix 2004, p. 46 eq. 3.3, p. 52 eq. 3.8.
}
