% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_out}
\alias{rad_sw_out}
\alias{rad_sw_out.default}
\alias{rad_sw_out.weather_station}
\title{Shortwave outgoing radiation}
\usage{
rad_sw_out(...)

\method{rad_sw_out}{default}(datetime, lon, lat, elev, temp, ..., surface_type = "field")

\method{rad_sw_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{weather_station}{Object of class weather_station.}
}
\value{
W/m\eqn{^2}.
}
\description{
Provide \code{slope} and \code{exposition} to perform topographic correction.
}
\references{
p46eq3.3, p52eq3.8
}
