% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_out}
\alias{rad_sw_out}
\alias{rad_sw_out.numeric}
\alias{rad_sw_out.weather_station}
\title{Reflected shortwave radiation}
\usage{
rad_sw_out(...)

\method{rad_sw_out}{numeric}(rad_sw_in, surface_type = "field", ...)

\method{rad_sw_out}{weather_station}(weather_station, surface_type = "field", ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{rad_sw_in}{Shortwave radiation on the ground onto a horizontal area in W/m².}

\item{surface_type}{type of surface for which an albedo will be selected.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Reflected shortwave radiation in W/m².
}
\description{
Calculation of the reflected shortwave radiation.
}
