% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_out}
\alias{rad_sw_out}
\alias{rad_sw_out.numeric}
\alias{rad_sw_out.weather_station}
\title{Reflected shortwave radiation}
\usage{
rad_sw_out(...)

\method{rad_sw_out}{numeric}(rad_sw_in, albedo, ...)

\method{rad_sw_out}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{rad_sw_in}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{albedo}{Albedo of the surface.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Reflected shortwave radiation in W/m^2.
}
\description{
Calculation of the reflected shortwave radiation.
}
