% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiance.R
\name{rad_sw_radiation_balance}
\alias{rad_sw_radiation_balance}
\title{Shortwave radiation balance}
\usage{
rad_sw_radiation_balance(rad_sw_ground_horizontal, rad_sw_reflected)
}
\arguments{
\item{rad_sw_ground_horizontal}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{rad_sw_reflected}{Reflected shortwave radiance in W/m^2.}
}
\value{
Shortwave radiation balance in W/m^2.
}
\description{
Calculation of the shortwave radiation balance.
}
