% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_reflected}
\alias{rad_sw_reflected}
\title{Reflected shortwave radiance}
\usage{
rad_sw_reflected(rad_sw_ground_horizontal, albedo)
}
\arguments{
\item{rad_sw_ground_horizontal}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{albedo}{Albedo of the surface.}
}
\value{
Reflected shortwave radiance in W/m^2.
}
\description{
Calculation of the reflected shortwave radiance.
}
