% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_reflected_by_terrain}
\alias{rad_sw_reflected_by_terrain}
\title{Shortwave radiation balance in dependency of topography}
\usage{
rad_sw_reflected_by_terrain(
  slope,
  valley = F,
  sol_elevation,
  sol_azimuth,
  exposition = 0,
  rad_sw_ground_horizontal,
  albedo,
  trans_total
)
}
\arguments{
\item{slope}{Slope in degrees.}

\item{valley}{Is the climate station placed in a valley (TRUE/FALSE)?}

\item{sol_elevation}{Sun elevation in degrees.}

\item{sol_azimuth}{Sun azimuth in degrees.}

\item{exposition}{Exposition (North = 0, South = 180)}

\item{rad_sw_ground_horizontal}{Shortwave radiation on the ground onto a horizontal area in W/m^2.}

\item{albedo}{Albedo of surface.}
}
\value{
Shortwave radiation balance in dependency of topography in W/m^2.
}
\description{
Calculate shortwave radiation balance in dependency of topography.
}
