% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_toa}
\alias{rad_sw_toa}
\alias{rad_sw_toa.default}
\alias{rad_sw_toa.weather_station}
\title{Shortwave radiation at top of atmosphere}
\usage{
rad_sw_toa(...)

\method{rad_sw_toa}{default}(datetime, lon, lat, ..., sol_const = sol_const_default)

\method{rad_sw_toa}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{sol_const}{Solar constant in W/m\eqn{^2}. Default = 1368.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
W/m\eqn{^2}.
}
\description{
Shortwave radiation at top of atmosphere
}
\references{
Bendix 2004, p. 244
}
