% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiation.R
\name{rad_sw_toa}
\alias{rad_sw_toa}
\alias{rad_sw_toa.default}
\alias{rad_sw_toa.weather_station}
\title{Shortwave radiation at top of atmosphere}
\usage{
rad_sw_toa(...)

\method{rad_sw_toa}{default}(datetime, lon, lat, ..., sol_const = 1368)

\method{rad_sw_toa}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{sol_const}{Solar constant in W/m\eqn{^2}}

\item{weather_station}{Object of class weather_station.}
}
\value{
Shortwave radiation at top of atmosphere in W/m\eqn{^2}.
}
\description{
Calculation of the shortwave radiation at the top of the atmosphere.
}
\references{
p244
}
