% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sensible_monin}
\alias{sensible_monin}
\alias{sensible_monin.numeric}
\alias{sensible_monin.weather_station}
\title{Sensible Heat using Monin-Obukhov length}
\usage{
sensible_monin(...)

\method{sensible_monin}{numeric}(
  t1,
  t2,
  p1,
  p2,
  z1 = 2,
  z2 = 10,
  monin,
  ustar,
  grad_rich_no,
  ...
)

\method{sensible_monin}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t1}{Air temperature at lower height in degrees C.}

\item{t2}{Air temperature at upper height in degrees C.}

\item{p1}{Pressure at lower height in hPa.}

\item{p2}{Pressure at upper height in hPa.}

\item{z1}{Lower height of measurement in m.}

\item{z2}{Upper height of measurement in m.}

\item{monin}{Monin-Obukhov-Length in m.}

\item{ustar}{Friction velocity in m/s.}

\item{grad_rich_no}{Gradient-Richardson-Number.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Sensible heat flux in W/m^2.
}
\description{
Calculates the sensible heat flux using the Monin-Obukhov length. Negative
flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
