% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sensible_priestley_taylor}
\alias{sensible_priestley_taylor}
\alias{sensible_priestley_taylor.default}
\alias{sensible_priestley_taylor.weather_station}
\title{Sensible Heat Priestley-Taylor Method}
\usage{
sensible_priestley_taylor(...)

\method{sensible_priestley_taylor}{default}(temp, rad_bal, soil_flux, surface_type, ...)

\method{sensible_priestley_taylor}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{temp}{Air temperature in °C.}

\item{rad_bal}{Radiation balance in W/m\eqn{^2}.}

\item{soil_flux}{Soil flux in W/m\eqn{^2}.}

\item{surface_type}{Surface type, for which a Priestley-Taylor coefficient will be selected. Options: field, bare soil, coniferous forest, water, wetland, spruce forest}
}
\value{
Sensible heat flux in W/m\eqn{^2}.
}
\description{
Calculates the Sensible heat flux using the Priestley-Taylor method. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
\references{
Foken 2016, p. 220, eq. 5.6
}
