% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensible.R
\name{sensible_priestley_taylor}
\alias{sensible_priestley_taylor}
\alias{sensible_priestley_taylor.default}
\alias{sensible_priestley_taylor.weather_station}
\title{Sensible Heat Priestley-Taylor Method}
\usage{
sensible_priestley_taylor(...)

\method{sensible_priestley_taylor}{default}(t, rad_bal, soil_flux, surface_type = "field", ...)

\method{sensible_priestley_taylor}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t}{Air temperature in °C.}

\item{rad_bal}{Radiation balance in W/m².}

\item{soil_flux}{Soil flux in W/m².}

\item{surface_type}{Surface type, for which a Priestley-Taylor coefficient will be selected. Default is for short grass.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Sensible heat flux in W/m².
}
\description{
Calculates the Sensible heat flux using the Priestley-Taylor method. Negative
heat flux signifies flux away from the surface, positive values signify flux
towards the surface.
}
\references{
Foken p220eq5.6.
}
