% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_attenuation}
\alias{soil_attenuation}
\alias{soil_attenuation.numeric}
\alias{soil_attenuation.weather_station}
\title{Soil attenuation length}
\usage{
soil_attenuation(...)

\method{soil_attenuation}{numeric}(moisture, texture = "sand", ...)

\method{soil_attenuation}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{moisture}{Soil moisture in Cubic meter/cubic meter}

\item{texture}{Soil texture. Either "sand", "peat" or "clay".}

\item{weather_station}{Object of class weather_station.}
}
\value{
Soil attenuation length in m.
}
\description{
Calculates soil attenuation length.
}
\references{
p253.
}
