% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_attenuation}
\alias{soil_attenuation}
\alias{soil_attenuation.default}
\alias{soil_attenuation.weather_station}
\title{Soil attenuation length}
\usage{
soil_attenuation(...)

\method{soil_attenuation}{default}(moisture, texture = "sand", ...)

\method{soil_attenuation}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{moisture}{Soil moisture in Cubic meter/cubic meter}

\item{texture}{Soil texture. Either "sand", "peat" or "clay".}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Soil attenuation length in m.
}
\description{
Calculates soil attenuation length.
}
\references{
Bendix 2004, p. 253.
}
