% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_attenuation}
\alias{soil_attenuation}
\title{Soil attenuation length}
\usage{
soil_attenuation(thermal_cond, vol_heat_cap)
}
\arguments{
\item{thermal_cond}{Thermal conductivity of soil (W/m K)}

\item{vol_heat_cap}{Volumetric heat capacity of soil \eq{J\emph{m^-3}K^-1}}
}
\value{
Soil attenuation length (m)
}
\description{
Soil attenuation length
}
