% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_attenuation}
\alias{soil_attenuation}
\title{Soil attenuation length}
\usage{
soil_attenuation(thermal_cond, vol_heat_cap)
}
\arguments{
\item{thermal_cond}{Thermal conductivity of soil in W/m K.}

\item{vol_heat_cap}{Volumetric heat capacity of soil in J/(m^3 * K).}
}
\value{
Soil attenuation length in m.
}
\description{
Calculates soil attenuation length.
}
