% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_cap}
\alias{soil_heat_cap}
\alias{soil_heat_cap.default}
\alias{soil_heat_cap.weather_station}
\title{Soil volumetric heat capacity}
\usage{
soil_heat_cap(...)

\method{soil_heat_cap}{default}(moisture, texture = "sand", ...)

\method{soil_heat_cap}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{moisture}{Soil moisture in Cubic meter/cubic meter}

\item{texture}{Soil texture. Either "sand", "peat" or "clay".}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Numeric vector with volumetric heat capacity in  MJ/ (m\eqn{^3} * K).
}
\description{
Calculates soil volumetric heat capacity (MJ / (m\eqn{^3} * K)) from soil moisture (Cubic meter/cubic meter) and texture.
}
\details{
Works by linearly interpolating volumetric heat capacity based on measured data.
}
\references{
Bendix 2004, p. 254.
}
