% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_cap}
\alias{soil_heat_cap}
\title{Soil volumetric heat capacity}
\usage{
soil_heat_cap(moisture, texture = "sand")
}
\arguments{
\item{moisture}{Soil moisture in Vol-\%.}

\item{texture}{Soil texture. Either "sand" or "clay".}
}
\value{
Numeric vector with volumetric heat capacity in J/(m^3 * K)
}
\description{
Calculates soil volumetric heat capacity (J / (m^3 * K)) from soil moisture (Vol-\%) and texture.
}
\details{
Works by linearly interpolating volumetric heat capacity based on measured data.
}
