% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_cap}
\alias{soil_heat_cap}
\alias{soil_heat_cap.numeric}
\alias{soil_heat_cap.weather_station}
\title{Soil volumetric heat capacity}
\usage{
soil_heat_cap(...)

\method{soil_heat_cap}{numeric}(moisture, texture = "sand", ...)

\method{soil_heat_cap}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{moisture}{Soil moisture in Vol-\%.}

\item{texture}{Soil texture. Either "sand" or "clay".}

\item{weather_station}{Object of class weather_station.}
}
\value{
Numeric vector with volumetric heat capacity in J/(m^3 * K)
}
\description{
Calculates soil volumetric heat capacity (J / (m^3 * K)) from soil moisture (Vol-\%) and texture.
}
\details{
Works by linearly interpolating volumetric heat capacity based on measured data.
}
