% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\alias{soil_heat_flux.default}
\alias{soil_heat_flux.weather_station}
\title{Soil heat flux}
\usage{
soil_heat_flux(...)

\method{soil_heat_flux}{default}(
  texture,
  moisture,
  soil_temp1,
  soil_temp2,
  soil_depth1,
  soil_depth2,
  ...
)

\method{soil_heat_flux}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=soil_thermal_cond.default]{soil_thermal_cond.default}}
  \describe{
    \item{\code{texture}}{Soil texture. Either \code{sand}, \code{peat} or \code{clay}.}
    \item{\code{moisture}}{Soil moisture in cubic meter/cubic meter.}
  }}

\item{soil_temp1, }{soil_temp2 Soil temperature in °C.}

\item{soil_depth, }{soil_depth2 Depth of the soil temperature measurement in m.}
}
\value{
Soil heat flux in W * m^(-2).
}
\description{
Calculates soil heat flux from measurements in two different
depths and thermal conductivity of the soil.
}
\details{
Negative values signify flux towards the atmosphere, positive values signify flux into the soil.
}
\references{
Bendix 2004, p. 71eq4.2.
}
