% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\alias{soil_heat_flux.default}
\alias{soil_heat_flux.weather_station}
\title{Soil heat flux}
\usage{
soil_heat_flux(...)

\method{soil_heat_flux}{default}(soil_temp, soil_depth, ...)

\method{soil_heat_flux}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=soil_thermal_cond.default]{soil_thermal_cond.default}}
  \describe{
    \item{\code{texture}}{Soil texture. Either \code{sand} (default), \code{peat} or \code{clay}.}
    \item{\code{moisture}}{Soil moisture in Cubic meter/cubic meter}
  }}

\item{soil_temp}{Soil temperature in °C. A vector with two elements.}

\item{soil_depth}{Depth of the measurement in m. A vector with two elements.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Soil heat flux in W * m^(-2).
}
\description{
Calculates soil heat flux from measurements in two different
depths and thermal conductivity of the soil.
}
\details{
Negative values signify flux towards the atmosphere, positive values signify flux into the soil.
}
\references{
p71eq4.2.
}
