% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\alias{soil_heat_flux.default}
\alias{soil_heat_flux.weather_station}
\title{Soil heat flux}
\usage{
soil_heat_flux(...)

\method{soil_heat_flux}{default}(
  texture,
  moisture,
  soil_temp1,
  soil_temp2,
  soil_depth1,
  soil_depth2,
  ...
)

\method{soil_heat_flux}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{texture}{Soil texture. Either "sand", "clay", or "peat".}

\item{moisture}{Soil moisture in cubic meter/cubic meter.}

\item{soil_temp1}{Soil temperature in degree Celcius of measurement 1.}

\item{soil_temp2}{Soil temperature in degree Celcius of measurement 2.}

\item{soil_depth1}{Depth of the soil temperature measurement 1 in m.}

\item{soil_depth2}{Depth of the soil temperature measurement 2 in m.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
\eqn{W \cdot m^-2}.
}
\description{
Calculates soil heat flux from measurements in two different
depths and thermal conductivity of the soil.
}
\details{
Negative values signify flux towards the atmosphere, positive values signify flux into the soil.
}
\references{
Bendix 2004, p. 71 eq. 4.2.
}
