% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_heat_flux}
\alias{soil_heat_flux}
\alias{soil_heat_flux.numeric}
\alias{soil_heat_flux.weather_station}
\title{Soil heat flux}
\usage{
soil_heat_flux(...)

\method{soil_heat_flux}{numeric}(ts1, ts2, depth1, depth2, thermal_cond, ...)

\method{soil_heat_flux}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{ts1}{Upper soil temperature (closest to the surface) in degrees C.}

\item{ts2}{Lower soil temperature in degrees C.}

\item{depth1}{Depth of upper measurement (closest to the surface) in m.}

\item{depth2}{Depth of lower measurement in m.}

\item{thermal_cond}{Thermal conductivity of soil in W/m K.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Soil heat flux in W*m^-2.
}
\description{
Calculates soil heat flux from measurements in two different
depths and thermal conductivity of the soil.
}
\details{
Negative values signify flux towards the atmosphere, positive values signify flux into the soil.
}
