% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_thermal_cond}
\alias{soil_thermal_cond}
\alias{soil_thermal_cond.default}
\alias{soil_thermal_cond.weather_station}
\title{Soil thermal conductivity}
\usage{
soil_thermal_cond(...)

\method{soil_thermal_cond}{default}(texture = "sand", moisture = 0, ...)

\method{soil_thermal_cond}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{texture}{Soil texture. Either \code{sand} (default), \code{peat} or \code{clay}.}

\item{moisture}{Soil moisture in Cubic meter/cubic meter}

\item{weather_station}{Object of class weather_station.}
}
\value{
Soil thermal conductivity in W/m K.
}
\description{
Calculates soil thermal conductivity (W/m K) from soil moisture (Cubic meter/cubic meter) and texture.
}
\details{
Works by linearly interpolating thermal conductivity based on measured data.
}
\references{
p254.
}
