% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_thermal_cond}
\alias{soil_thermal_cond}
\title{Soil thermal conductivity}
\usage{
soil_thermal_cond(moisture, texture = "sand")
}
\arguments{
\item{moisture}{Soil moisture in Vol-\%.}

\item{texture}{Soil texture. Either "sand" or "clay".}
}
\value{
Soil thermal conductivity in W/m K.
}
\description{
Calculates soil thermal conductivity (W/m K) from soil moisture (Vol-\%) and texture.
}
\details{
Works by linearly interpolating thermal conductivity based on measured data.
}
