% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\name{soil_thermal_cond}
\alias{soil_thermal_cond}
\alias{soil_thermal_cond.numeric}
\alias{soil_thermal_cond.weather_station}
\title{Soil thermal conductivity}
\usage{
soil_thermal_cond(...)

\method{soil_thermal_cond}{numeric}(moisture, texture = "sand", ...)

\method{soil_thermal_cond}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{moisture}{Soil moisture in Vol-\%.}

\item{texture}{Soil texture. Either "sand" or "clay".}

\item{weather_station}{Object of class weather_station.}
}
\value{
Soil thermal conductivity in W/m K.
}
\description{
Calculates soil thermal conductivity (W/m K) from soil moisture (Vol-\%) and texture.
}
\details{
Works by linearly interpolating thermal conductivity based on measured data.
}
