% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_azimuth}
\alias{sol_azimuth}
\alias{sol_azimuth.default}
\alias{sol_azimuth.weather_station}
\title{Solar azimuth angle}
\usage{
sol_azimuth(...)

\method{sol_azimuth}{default}(datetime, lon, lat, ...)

\method{sol_azimuth}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Solar azimuth angle in degrees.
}
\description{
Calculates solar azimuth angle for the given date and time.
}
\references{
p243
}
