% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_azimuth}
\alias{sol_azimuth}
\alias{sol_azimuth.default}
\alias{sol_azimuth.weather_station}
\title{Solar azimuth}
\usage{
sol_azimuth(...)

\method{sol_azimuth}{default}(datetime, lon, lat, ...)

\method{sol_azimuth}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
degree
}
\description{
Solar azimuth
}
\references{
p243
}
