% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_day_angle}
\alias{sol_day_angle}
\alias{sol_day_angle.POSIXt}
\title{Day angle}
\usage{
sol_day_angle(...)

\method{sol_day_angle}{POSIXt}(datetime, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}
}
\value{
degree
}
\description{
Day angle
}
\references{
p243
}
