% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_eccentricity}
\alias{sol_eccentricity}
\alias{sol_eccentricity.weather_station}
\title{Eccentricity factor}
\usage{
sol_eccentricity(...)

\method{sol_eccentricity}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Unitless.
}
\description{
The track of Earth around Sun is not a circle, but more like an ellipse.
}
