% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_eccentricity}
\alias{sol_eccentricity}
\title{Eccentricity}
\usage{
sol_eccentricity(datetime)
}
\arguments{
\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \link[base]{as.Posixlt} and \link[base]{strptime} for conversion.}
}
\value{
Eccentricity at the date.
}
\description{
Calculates the eccentricity.
}
