% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_ecliptic_length}
\alias{sol_ecliptic_length}
\alias{sol_ecliptic_length.default}
\alias{sol_ecliptic_length.weather_station}
\title{Solar ecliptic length}
\usage{
sol_ecliptic_length(...)

\method{sol_ecliptic_length}{default}(datetime, ...)

\method{sol_ecliptic_length}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
degree
}
\description{
Solar ecliptic length
}
\references{
Bendix 2004, p. 243
}
