% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_elevation}
\alias{sol_elevation}
\alias{sol_elevation.default}
\alias{sol_elevation.weather_station}
\title{Solar elevation angle}
\usage{
sol_elevation(...)

\method{sol_elevation}{default}(datetime, lon, lat, ...)

\method{sol_elevation}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Solar elevation angle in degrees.

degree
}
\description{
Calculates solar elevation angle for the given date and time.
}
\references{
p243
}
