% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_elevation}
\alias{sol_elevation}
\alias{sol_elevation.default}
\alias{sol_elevation.weather_station}
\title{Solar elevation angle}
\usage{
sol_elevation(...)

\method{sol_elevation}{default}(datetime, lon, lat, ...)

\method{sol_elevation}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.
Make sure to provide the correct timezone information!}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Degree.
}
\description{
Solar elevation angle
}
\references{
Bendix 2004, p. 243.
}
