% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_elevation}
\alias{sol_elevation}
\alias{sol_elevation.POSIXt}
\alias{sol_elevation.weather_station}
\title{Solar elevation angle}
\usage{
sol_elevation(...)

\method{sol_elevation}{POSIXt}(datetime, lat, lon, ...)

\method{sol_elevation}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.}

\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Solar elevation angle in degrees.
}
\description{
Calculates solar elevation angle for the given date and time.
}
