% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_hour_angle}
\alias{sol_hour_angle}
\alias{sol_hour_angle.default}
\alias{sol_hour_angle.weather_station}
\title{Solar hour angle}
\usage{
sol_hour_angle(...)

\method{sol_hour_angle}{default}(datetime, lon, ...)

\method{sol_hour_angle}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{weather_station}{Object of class weather_station.}
}
\value{
degree
}
\description{
Solar hour angle
}
\references{
p243
}
