% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_hour_angle}
\alias{sol_hour_angle}
\alias{sol_hour_angle.default}
\alias{sol_hour_angle.weather_station}
\title{Solar hour angle}
\usage{
sol_hour_angle(...)

\method{sol_hour_angle}{default}(datetime, lon, ...)

\method{sol_hour_angle}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
degree
}
\description{
Solar hour angle
}
\references{
Bendix 2004, p. 243
}
