% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_julian_day}
\alias{sol_julian_day}
\alias{sol_julian_day.default}
\title{Julian day}
\usage{
sol_julian_day(...)

\method{sol_julian_day}{default}(datetime, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}
}
\value{
unitless
}
\description{
Julian day
}
