% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_julian_day}
\alias{sol_julian_day}
\alias{sol_julian_day.POSIXt}
\title{Julian day}
\usage{
sol_julian_day(...)

\method{sol_julian_day}{POSIXt}(datetime, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{datetime in POSIXt}
}
\value{
unitless
}
\description{
Julian day
}
\references{
p
}
