% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_julian_day}
\alias{sol_julian_day}
\title{Julian day}
\usage{
sol_julian_day(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
Unitless.
}
\description{
Day of year as an integer from 1 to 366.
}
