% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_julian_day}
\alias{sol_julian_day}
\alias{sol_julian_day.default}
\alias{sol_julian_day.weather_station}
\title{Julian day}
\usage{
sol_julian_day(...)

\method{sol_julian_day}{default}(datetime, ...)

\method{sol_julian_day}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
unitless
}
\description{
Day of year as an integer from 1 to 366.
}
