% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_medium_anomaly}
\alias{sol_medium_anomaly}
\alias{sol_medium_anomaly.default}
\alias{sol_time_formula.default}
\title{Solar medium anomaly}
\usage{
sol_medium_anomaly(...)

\method{sol_medium_anomaly}{default}(datetime, ...)

\method{sol_time_formula}{default}(datetime, lon, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}
}
\value{
degree
}
\description{
Solar medium anomaly
}
\references{
p243

p243
}
