% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{sol_medium_suntime}
\alias{sol_medium_suntime}
\alias{sol_medium_suntime.default}
\alias{sol_medium_suntime.weather_station}
\title{Solar medium suntime}
\usage{
sol_medium_suntime(...)

\method{sol_medium_suntime}{default}(datetime, lon, ...)

\method{sol_medium_suntime}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{weather_station}{Object of class weather_station.}
}
\value{
hour
}
\description{
Solar medium suntime
}
\references{
p243
}
