% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surface_properties}
\alias{surface_properties}
\title{Emissivity and roughness lenght for different surfaces.}
\format{
A data frame with 10 rows and 3 variables:
\describe{
\item{surface_type}{Type of surface}
\item{emissivity}{Emissivity of the chosen surface}
\item{roughness_length}{Rougness lenght of surface in }
}
}
\usage{
surface_properties
}
\description{
A dataset containing values for emissivity and roughness length for different types of surface
}
\keyword{datasets}
