% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{temp_pot_temp}
\alias{temp_pot_temp}
\alias{temp_pot_temp.default}
\alias{temp_pot_temp.weather_station}
\title{Potential Temperature}
\usage{
temp_pot_temp(...)

\method{temp_pot_temp}{default}(t, elev, ...)

\method{temp_pot_temp}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{t}{Temperature in °C.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{Object of class \code{weather_station}.}

\item{height}{Height of measurement, either "upper" or "lower".}
}
\value{
Potential temperature in °C.
}
\description{
Calculation of the potential air temperature.
}
\references{
Bendix 2004, p. 261.
}
