% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{temp_pot_temp}
\alias{temp_pot_temp}
\title{Potential Temperature}
\usage{
temp_pot_temp(t, p)
}
\arguments{
\item{p}{Pressure in hPa.}

\item{t1}{Temperature in degrees C.}
}
\value{
Potential temperature in K.
}
\description{
Calculation of the potential air temperature.
}
