% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terr_sky_view}
\alias{terr_sky_view}
\alias{terr_sky_view.default}
\alias{terr_sky_view.weather_station}
\title{Sky view factor}
\usage{
terr_sky_view(...)

\method{terr_sky_view}{default}(slope, valley, ...)

\method{terr_sky_view}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{slope}{Slope in degree.}

\item{valley}{Is the position in a valley (\code{TRUE}) or on a slope (\code{FALSE})?}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
How much sky can be seen from the point?
}
\details{
Terrain view factor can be calculated by \code{1 - terr_sky_view()}, which means:
"How much sky can \strong{not} be seen from the point?" or
"How much terrain can be seen from the point?"
}
\references{
Bendix 2004, p. 63 eq. 3.15.

Bendix 2004, p. 57 eq. 3.12, p. 57 eq. 3.13.
}
