% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terr_sky_view}
\alias{terr_sky_view}
\alias{terr_sky_view.numeric}
\alias{terr_sky_view.weather_station}
\title{Sky view factor}
\usage{
terr_sky_view(...)

\method{terr_sky_view}{numeric}(slope, valley = F, ...)

\method{terr_sky_view}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{slope}{Inclination of slope in degrees.}

\item{valley}{If the position is in a valley (TRUE) or on a slope (FALSE).}

\item{weather_station}{Object of class weather_station.}
}
\value{
Sky view factor from 0-1.
}
\description{
Calculates the sky view factor of a position only based on slope angle and
position in valley or on a slope.
}
\details{
Terrain view factor can be calculated by 1-terr_sky_view.
}
