% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terr_terrain_angle}
\alias{terr_terrain_angle}
\alias{terr_terrain_angle.default}
\alias{terr_terrain_angle.weather_station}
\title{Terrain angle}
\usage{
terr_terrain_angle(...)

\method{terr_terrain_angle}{default}(datetime, lon, lat, slope, exposition, ...)

\method{terr_terrain_angle}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{slope}{Slope in degree.}

\item{exposition}{Exposition in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Degree.
}
\description{
Terrain angle
}
\references{
Bendix 2004, p. 52 eq. 3.7.
}
