% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_aerosol}
\alias{trans_aerosol}
\alias{trans_aerosol.default}
\alias{trans_aerosol.weather_station}
\title{Transmittance due to aerosols}
\usage{
trans_aerosol(...)

\method{trans_aerosol}{default}(datetime, lon, lat, elev, temp, ..., vis = vis_default)

\method{trans_aerosol}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}. See \code{\link[base:as.POSIXlt]{base::as.POSIXlt()}}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in \eqn{^\circ}C.}

\item{vis}{Visibility in km, default 30.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
Transmittance due to fine particles in the air.
}
\details{
Visibility is used to linearly interpolate aerosol optical thickness.
}
\references{
Bendix 2004, p. 246.
}
