% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_aerosol}
\alias{trans_aerosol}
\title{Transmittance due to aerosols}
\usage{
trans_aerosol(air_mass_abs, vis)
}
\arguments{
\item{vis}{}
}
\value{

}
\description{
Transmittance due to aerosols
}
