% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_aerosol}
\alias{trans_aerosol}
\alias{trans_aerosol.numeric}
\alias{trans_aerosol.weather_station}
\title{Transmittance due to aerosols}
\usage{
trans_aerosol(...)

\method{trans_aerosol}{numeric}(air_mass_abs, vis = 30, ...)

\method{trans_aerosol}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{air_mass_abs}{Absolute optical air mass.}

\item{vis}{Visibility in km.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Transmittance due to aerosols (0-1).
}
\description{
Calculates transmittance due to aerosols.
}
