% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_abs}
\alias{trans_air_mass_abs}
\alias{trans_air_mass_abs.default}
\alias{trans_air_mass_abs.weather_station}
\title{Absolute optical air mass}
\usage{
trans_air_mass_abs(...)

\method{trans_air_mass_abs}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_air_mass_abs}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=pres_p.default]{pres_p.default}}
  \describe{
    \item{\code{g}}{Gravitational acceleration in m/s\eqn{^2}, default 9.81.}
    \item{\code{rl}}{Specific gas constant for air in m\eqn{^2}/s\eqn{^2}/K, default 287.05.}
  }}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Absolute optical air mass.
}
\description{
Calculates absolute optical air mass.
}
\references{
p247.
}
