% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_abs}
\alias{trans_air_mass_abs}
\alias{trans_air_mass_abs.default}
\alias{trans_air_mass_abs.weather_station}
\title{Absolute optical air mass}
\usage{
trans_air_mass_abs(...)

\method{trans_air_mass_abs}{default}(datetime, lon, lat, elev, temp, p0 = 1013, ...)

\method{trans_air_mass_abs}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Absolute optical air mass.

unitless
}
\description{
Calculates absolute optical air mass.
}
\references{
p247.
}
