% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_abs}
\alias{trans_air_mass_abs}
\alias{trans_air_mass_abs.numeric}
\alias{trans_air_mass_abs.weather_station}
\title{Absolute optical air mass}
\usage{
trans_air_mass_abs(...)

\method{trans_air_mass_abs}{numeric}(lat, datetime, lon, elev, t, p0 = 1013)

\method{trans_air_mass_abs}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{lat}{Latitude in decimal degrees.}

\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}

\item{air_mass_rel}{Relative optical air mass.}

\item{p}{Air pressure in hPa.}
}
\value{
Absolute optical air mass.

unitless
}
\description{
Calculates absolute optical air mass.
}
\references{
p247.
}
