% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_abs}
\alias{trans_air_mass_abs}
\title{Absolute optical air mass}
\usage{
trans_air_mass_abs(air_mass_rel, pressure)
}
\arguments{
\item{pressure}{}
}
\value{

}
\description{
Absolute optical air mass
}
