% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_rel}
\alias{trans_air_mass_rel}
\alias{trans_air_mass_rel.default}
\alias{trans_air_mass_rel.weather_station}
\title{Relative optical air mass}
\usage{
trans_air_mass_rel(...)

\method{trans_air_mass_rel}{default}(datetime, lon, lat, ...)

\method{trans_air_mass_rel}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
unitless
}
\description{
Relative optical air mass
}
\references{
p246.
}
