% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_rel}
\alias{trans_air_mass_rel}
\alias{trans_air_mass_rel.default}
\alias{trans_air_mass_rel.weather_station}
\title{Relative optical air mass}
\usage{
trans_air_mass_rel(...)

\method{trans_air_mass_rel}{default}(datetime, lon, lat, ...)

\method{trans_air_mass_rel}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Relative optical air mass.

unitless
}
\description{
Calculates relative optical air mass. Returns NA for negative values.
}
\references{
p246.
}
