% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_air_mass_rel}
\alias{trans_air_mass_rel}
\title{Relative optical air mass}
\usage{
trans_air_mass_rel(sol_elevation)
}
\arguments{
\item{sol_elevation}{Solar elevation in degrees.}
}
\value{
Relative optical air mass.
}
\description{
Calculates relative optical air mass.
}
