% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_gas}
\alias{trans_gas}
\alias{trans_gas.numeric}
\alias{trans_gas.weather_station}
\title{Transmittance due to gas}
\usage{
trans_gas(...)

\method{trans_gas}{numeric}(lat, datetime, lon, elev, temp)

\method{trans_gas}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Transmittance due to gas (0-1), unitless

unitless
}
\description{
Calculates transmittance due to O$\emph{{2}$ and CO$}{2}$.
}
\references{
p246.
}
