% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_gas}
\alias{trans_gas}
\alias{trans_gas.numeric}
\alias{trans_gas.weather_station}
\title{Transmittance due to gas}
\usage{
trans_gas(...)

\method{trans_gas}{numeric}(lat, datetime, lon, elev, t)

\method{trans_gas}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{lat}{Latitude in decimal degrees.}

\item{datetime}{POSIXt object (POSIXct, POSIXlt).
See \link[base:as.POSIXlt]{base::as.POSIXlt} and \link[base:strptime]{base::strptime} for conversion.}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}

\item{air_mass_abs}{Absolute optical air mass.}
}
\value{
Transmittance due to gas (0-1)

unitless
}
\description{
Calculates transmittance due to O$\emph{{2}$ and CO$}{2}$.
}
\references{
p246.
}
