% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_gas}
\alias{trans_gas}
\alias{trans_gas.default}
\alias{trans_gas.weather_station}
\title{Transmittance due to gas}
\usage{
trans_gas(...)

\method{trans_gas}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_gas}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime of class \code{POSIXlt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
Calculates transmittance due to O\eqn{_2} and CO\eqn{_2}.
}
\references{
p246.
}
