% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_gas}
\alias{trans_gas}
\alias{trans_gas.default}
\alias{trans_gas.weather_station}
\title{Transmittance due to gas}
\usage{
trans_gas(...)

\method{trans_gas}{default}(datetime, lon, lat, elev, temp, ..., p0 = 1013)

\method{trans_gas}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude.}

\item{lat}{Latitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in °C.}

\item{p0}{Standard pressure in hPa.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Transmittance due to gas (0-1), unitless

unitless
}
\description{
Calculates transmittance due to O$\emph{{2}$ and CO$}{2}$.
}
\references{
p246.
}
