% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_ozone}
\alias{trans_ozone}
\alias{trans_ozone.default}
\alias{trans_ozone.weather_station}
\title{Transmittance due to ozone}
\usage{
trans_ozone(...)

\method{trans_ozone}{default}(datetime, lon, lat, ..., ozone_column = ozone_column_default)

\method{trans_ozone}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{ozone_column}{Atmospheric ozone as column in cm, default 0.35.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
Transmittance due to ozone
}
\references{
p245.
}
