% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_rayleigh}
\alias{trans_rayleigh}
\alias{trans_rayleigh.default}
\alias{trans_rayleigh.weather_station}
\title{Transmittance due to rayleigh scattering}
\usage{
trans_rayleigh(...)

\method{trans_rayleigh}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_rayleigh}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{elev}{Elevation above sea level in m.}

\item{temp}{Temperature in degree Celcius.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
Transmittance due to rayleigh scattering
}
\references{
p245.
}
