% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_rayleigh}
\alias{trans_rayleigh}
\alias{trans_rayleigh.default}
\alias{trans_rayleigh.weather_station}
\title{Transmittance due to rayleigh scattering}
\usage{
trans_rayleigh(...)

\method{trans_rayleigh}{default}(datetime, lon, lat, elev, temp, ..., p0 = 1013)

\method{trans_rayleigh}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude in decimal degrees.}

\item{elev}{Elevation above sea level in m.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Transmittance due to rayleigh scattering (0-1). unitless
}
\description{
Calculates transmittance due to rayleigh scattering.
}
\references{
p245.
}
