% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_rayleigh}
\alias{trans_rayleigh}
\alias{trans_rayleigh.numeric}
\alias{trans_rayleigh.weather_station}
\title{Transmittance due to rayleigh scattering}
\usage{
trans_rayleigh(...)

\method{trans_rayleigh}{numeric}(lat, datetime, lon, elev, t)

\method{trans_rayleigh}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}

\item{air_mass_abs}{Absolute optical air mass.}
}
\value{
Transmittance due to rayleigh scattering (0-1).

unitless
}
\description{
Calculates transmittance due to rayleigh scattering.
}
\references{
p245.
}
