% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_rayleigh}
\alias{trans_rayleigh}
\alias{trans_rayleigh.numeric}
\alias{trans_rayleigh.weather_station}
\title{Transmittance due to rayleigh scattering}
\usage{
trans_rayleigh(...)

\method{trans_rayleigh}{numeric}(air_mass_abs, ...)

\method{trans_rayleigh}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{air_mass_abs}{Absolute optical air mass.}

\item{weather_station}{Object of class weather_station}
}
\value{
Transmittance due to rayleigh scattering (0-1).
}
\description{
Calculates transmittance due to rayleigh scattering.
}
