% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_total}
\alias{trans_total}
\title{Total transmittance}
\usage{
trans_total(sol_elevation, t, elev, oz = 0.35, vis = 30, p = NULL)
}
\arguments{
\item{sol_elevation}{Solar elevation in degrees.}

\item{t}{Air temperature in degrees C.}

\item{elev}{Altitude above sea level in m.}

\item{oz}{Columnar ozone in cm. Default is average global value.}

\item{vis}{Meteorological visibility in km. Default is the visibility on a clear day.}

\item{p}{OPTIONAL. Pressure in hPa. Estimated from elev and t if not available.}
}
\value{
Total transmittance (0-1)
}
\description{
Calculates total transmittance of the atmosphere.
}
