% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_total}
\alias{trans_total}
\title{Total transmittance}
\usage{
trans_total(
  sol_elevation,
  temp,
  altitude,
  oz = 0.35,
  vis = 30,
  pressure = NULL
)
}
\arguments{
\item{pressure}{}
}
\value{

}
\description{
Total transmittance
}
