% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_vapor}
\alias{trans_vapor}
\alias{trans_vapor.numeric}
\alias{trans_vapor.weather_station}
\title{Transmittance due to water vapor}
\usage{
trans_vapor(...)

\method{trans_vapor}{numeric}(lat, datetime, lon, elev, t)

\method{trans_vapor}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{datetime}{Datetime in POSIXt}

\item{lon}{Longitude in decimal degrees.}

\item{weather_station}{Object of class weather_station.}

\item{air_mass_rel}{Relative optical air mass.}

\item{precipitable_water}{Precipitable water in cm.}
}
\value{
Transmittance due to water vapor (0-1).

unitless
}
\description{
Calculates transmittance due to water vapor.
}
\references{
p245.
}
