% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_vapor}
\alias{trans_vapor}
\alias{trans_vapor.default}
\alias{trans_vapor.weather_station}
\title{Transmittance due to water vapor}
\usage{
trans_vapor(...)

\method{trans_vapor}{default}(datetime, lon, lat, elev, temp, ...)

\method{trans_vapor}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{datetime}{Datetime in \code{POSIXt}.}

\item{lon}{Longitude in degree.}

\item{lat}{Latitude in degree.}

\item{weather_station}{Object of class \code{weather_station}.}
}
\value{
Ratio from 0 to 1, unitless.
}
\description{
Transmittance due to water vapor
}
\references{
p245.
}
