% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittance.R
\name{trans_vapor}
\alias{trans_vapor}
\title{Transmittance due to water vapor}
\usage{
trans_vapor(air_mass_rel, pw)
}
\arguments{
\item{air_mass_rel}{Relative optical air mass.}

\item{pw}{Precipitable water in cm.}
}
\value{
Transmittance due to water vapor (0-1).
}
\description{
Calculates transmittance due to water vapor.
}
