% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_ex_quotient_temp}
\alias{turb_flux_ex_quotient_temp}
\alias{turb_flux_ex_quotient_temp.numeric}
\alias{turb_flux_ex_quotient_temp.weather_station}
\title{Exchange quotient for heat transmission}
\usage{
turb_flux_ex_quotient_temp(...)

\method{turb_flux_ex_quotient_temp}{numeric}(
  t1,
  t2,
  z1 = 2,
  z2 = 10,
  v1,
  v2,
  elev,
  surface_type,
  ...
)

\method{turb_flux_ex_quotient_temp}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in °C.}

\item{t2}{Temperature at upper height in degrees C.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{elev}{Elevation above sea level in m.}

\item{surface_type}{Type of surface.}

\item{weather_station}{Object of class weather_station}
}
\value{
Exchange quotient for heat transmission in kg/(m*s).
}
\description{
Calculation of the exchange quotient of the turbulent heat transmission.
}
\references{
Foken p362 Businger.
}
