% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_ex_quotient_temp}
\alias{turb_flux_ex_quotient_temp}
\title{Exchange quotient for heat transmission.}
\usage{
turb_flux_ex_quotient_temp(stability, ustar, mol, z1, air_density)
}
\arguments{
\item{stability}{Stability of atmosphere. One of "stabil", "neutral" or "instabil".}

\item{ustar}{Friction velocity in m/s.}

\item{mol}{Monin-Obhukov-Length in m.}

\item{z1}{Height in m.}

\item{air_density}{Air density in kg*m^(-3).}
}
\value{
Exchange quotient for heat transmission in kg/(m*s).
}
\description{
Exchange quotient for heat transmission.
}
