% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_ex_quotient_temp}
\alias{turb_flux_ex_quotient_temp}
\alias{turb_flux_ex_quotient_temp.numeric}
\alias{turb_flux_ex_quotient_temp.weather_station}
\title{Exchange quotient for heat transmission}
\usage{
turb_flux_ex_quotient_temp(...)

\method{turb_flux_ex_quotient_temp}{numeric}(grad_rich_no, ustar, monin, z, air_density, ...)

\method{turb_flux_ex_quotient_temp}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{grad_rich_no}{Gradient-Richardson-Number.}

\item{ustar}{Friction velocity in m/s.}

\item{monin}{Monin-Obhukov-Length in m.}

\item{z}{Height in m.}

\item{air_density}{Air density in kg/m^3.}

\item{weather_station}{Object of class weather_station}

\item{height}{Height of measurement. Either "upper" or "lower".}
}
\value{
Exchange quotient for heat transmission in kg/(m*s).
}
\description{
Calculation of the exchange quotient of the turbulent heat transmission.
}
