% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_grad_rich_no}
\alias{turb_flux_grad_rich_no}
\alias{turb_flux_grad_rich_no.numeric}
\alias{turb_flux_grad_rich_no.weather_station}
\title{Gradient-Richardson-Number}
\usage{
turb_flux_grad_rich_no(...)

\method{turb_flux_grad_rich_no}{numeric}(t1, t2, z1 = 2, z2 = 10, v1, v2, p1, p2, ...)

\method{turb_flux_grad_rich_no}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in degrees C.}

\item{t2}{Temperature at upper height in degrees C.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{p1}{Pressure at lower height (e.g. height of anemometer) in hPa.}

\item{p2}{Pressure at upper height in hPa.}

\item{weather_station}{Object of class weather_station}
}
\value{
Gradient-Richardson-Number.
}
\description{
Calculation of the Gradient-Richardson-Number. The number represents the
stability of the atmosphere. Negative values signify unstable conditions,
positive values signify stable conditions.
}
