% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_imp_exchange}
\alias{turb_flux_imp_exchange}
\alias{turb_flux_imp_exchange.numeric}
\alias{turb_flux_imp_exchange.weather_station}
\title{Turbulent impulse exchange}
\usage{
turb_flux_imp_exchange(...)

\method{turb_flux_imp_exchange}{numeric}(ex_quotient, v1, v2, z1 = 2, z2 = 10, ...)

\method{turb_flux_imp_exchange}{weather_station}(weather_station, height = "lower", ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{ex_quotient}{Exchange quotient in kg/(m*s).}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{weather_station}{Object of class weather_station}

\item{height}{Height of measurement. Either "upper" or "lower".}
}
\value{
Turbulent impulse exchange in kg/(m*s^2).
}
\description{
Calculation of the turbulent impulse exchange.
}
