% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulent_flux.R
\name{turb_flux_monin}
\alias{turb_flux_monin}
\alias{turb_flux_monin.numeric}
\alias{turb_flux_monin.weather_station}
\title{Monin-Obhukov-Length}
\usage{
turb_flux_monin(...)

\method{turb_flux_monin}{numeric}(grad_rich_no, z1 = 2, z2 = 10, z0, v1, v2, t1, t2, ...)

\method{turb_flux_monin}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional parameters passed to later functions.}

\item{grad_rich_no}{Gradient-Richardson-Number.}

\item{z1}{Lower height of measurement (e.g. height of anemometer) in m.}

\item{z2}{Upper height of measurement in m.}

\item{z0}{Roughness length in m.}

\item{v1}{Windspeed at lower height (e.g. height of anemometer) in m/s.}

\item{v2}{Windspeed at upper height in m/s.}

\item{t1}{Temperature at lower height (e.g. height of anemometer) in degrees C.}

\item{t2}{Temperature at upper height in degrees C.}

\item{weather_station}{Object of class weather_station}
}
\value{
Monin-Obhukov-Length in m.
}
\description{
Calculation of the Monin-Obhukov-Length.
Stability of atmosphere needs to be given as one of "stabil", "neutral" or "unstabil".
}
