% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence.R
\name{turb_roughness_length}
\alias{turb_roughness_length}
\alias{turb_roughness_length.default}
\alias{turb_roughness_length.weather_station}
\title{Roughness length}
\usage{
turb_roughness_length(...)

\method{turb_roughness_length}{default}(surface_type = NULL, obs_height = NULL, ...)

\method{turb_roughness_length}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{surface_type}{Type of surface. Options: field, acre, lawn, street, agriculture, settlement, coniferous forest, deciduous forest, mixed forest, city, water, shrub}

\item{obs_height}{Height of obstacle in m.}
}
\value{
roughness length in m.
}
\description{
Calculate the roughness length of a surface on the basis of the obstacle height or the type of the surface.
Possible surface types are:
"field", "acre", "lawn", "street", "agriculture", "settlement", "coniferous forest", "deciduous forest", "mixed forest", "city"
You need to specify only one, "surface_type" OR "obs_height".
}
\references{
Bendix 2004, p. 239
}
