% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turbulence.R
\name{turb_ustar}
\alias{turb_ustar}
\alias{turb_ustar.default}
\alias{turb_ustar.weather_station}
\title{Friction velocity}
\usage{
turb_ustar(...)

\method{turb_ustar}{default}(v, z, ...)

\method{turb_ustar}{weather_station}(weather_station, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=turb_roughness_length]{turb_roughness_length}}
  \describe{
    \item{\code{}}{}
  }}

\item{v}{Windspeed in height of anemometer in m/s.}

\item{z}{Height of anemometer in m.}

\item{weather_station}{Object of class weather_station.}
}
\value{
Friction velocity in m/s.
}
\description{
Calculate the friction velocity of the surface.
}
\references{
Bendix 2004, p. 239
}
